% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_manipulation.R
\name{add_group_id}
\alias{add_group_id}
\title{Add column containing row number within groups}
\usage{
add_group_id(.df, ..., .arrange_by = NULL, .new_colname = "n")
}
\arguments{
\item{.df}{Data frame}

\item{...}{Columns used to form groups. Passed to \code{dplyr::group_by}}

\item{.arrange_by}{Columns used to order \code{.df} prior to grouping. Enter as character vector}

\item{.new_colname}{Name for newly created column containing group number. Enter as quoted or unquoted column name. Default is \code{"n"}.}
}
\value{
Ungrouped data frame with integer column (\code{.new_colname}) containing group number
}
\description{
Add column containing row number within groups
}
