% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Addins.R
\name{addin_reshape}
\alias{addin_reshape}
\title{Separate function arguments onto separate lines}
\usage{
addin_reshape(...)
}
\value{
Code silently edited
}
\description{
Functionality from Lionel Henry's excellent package codegrip
}
