% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectors.R
\name{age_group}
\alias{age_group}
\alias{decade}
\title{Convert numeric vector of ages to age groups}
\usage{
age_group(
  x,
  width = 10,
  include_lower = TRUE,
  include_upper = FALSE,
  min = 0,
  max = range(x, finite = TRUE)[2L]
)

decade(x, as_fct = TRUE)
}
\arguments{
\item{x}{Integer vector containing years}

\item{width}{Width of each age interval in years. Enter as length 1 numeric. Default is \code{10} (creates age deciles)}

\item{include_lower}{If \code{TRUE} (default), all values in a given age interval must be greater than or equal to the lower limit that defines that age interval. If \code{FALSE}, all values in a given age interval  must be greater than the lower limit that defines that age interval}

\item{include_upper}{If \code{TRUE}, all values in a given age interval  must be less than or equal to the upper limit that defines that age interval. If \code{FALSE} (default), all values in a given age interval  must be less than the upper limit that defines that age interval}

\item{min}{Lower limit of first age age interval. Enter as length 1 numeric. Default is \code{0}}

\item{max}{Upper limit of last age age interval. Enter as length 1 numeric. Default uses nearest multiple of 10 that is greater than or equal to the maximum value in \code{x}}

\item{as_fct}{If \code{TRUE} (default), output converted to a factor. If \code{FALSE}, output will be an integer}
}
\value{
Factor vector with same length as input

Factor (if \code{as_fct = TRUE}) or integer (if \code{as_fct = FALSE}) vector with same length as input indicating decade
}
\description{
Convert numeric vector of ages to age groups

Convert year to decade
}
