% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{all_equal}
\alias{all_equal}
\alias{all_identical}
\title{Recursively determine whether all inputs are equal}
\usage{
all_equal(..., tolerance = 1.5e-08)

all_identical(...)
}
\arguments{
\item{...}{Either a comma separated list of vectors not wrapped in \code{c()} or a list containing vectors}

\item{tolerance}{Threshold difference between values above which output will be \code{FALSE}. Default is \code{1.5e-8}}
}
\value{
Length 1 logical. \code{TRUE} if predicate met, \code{FALSE} if not
}
\description{
Recursively determine whether all inputs are equal

Recursively determine whether all inputs are identical
}
