% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectors.R
\name{allNA}
\alias{allNA}
\title{Check if vector contains only missing values}
\usage{
allNA(x)
}
\arguments{
\item{x}{Vector}
}
\value{
Length 1 logical vector
}
\description{
Check if vector contains only missing values
}
