% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{alpha}
\alias{alpha}
\title{Apply alpha filter to color}
\usage{
alpha(color, alpha = 1)
}
\arguments{
\item{color}{Character vector of color names or hexadecimal codes}

\item{alpha}{Opacity. Enter as numeric 0-1 (1 for completely opaque, 0 for completely transparent). Default is \code{0.8}}
}
\description{
Functionality from \code{grDevices::adjustcolor}
}
