% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectors.R
\name{any_dupes}
\alias{any_dupes}
\alias{all_unique}
\title{Determine whether vector contains any duplicate values}
\usage{
any_dupes(x, na.rm = TRUE)

all_unique(x, na.rm = TRUE)
}
\arguments{
\item{x}{Vector}

\item{na.rm}{If \code{TRUE} (default), missing values are removed prior to determining whether duplicates are present}
}
\value{
Length 1 logical vector
}
\description{
Determine whether vector contains any duplicate values

Determine whether all values in a vector are unique
}
