% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectors.R
\name{any_zero}
\alias{any_zero}
\title{Determine whether any values in a vector are equal to zero}
\usage{
any_zero(x, na.rm = TRUE)
}
\arguments{
\item{x}{Numeric vector}

\item{na.rm}{If \code{TRUE} (default), missing values removed. If \code{FALSE} and \code{x} contains \code{NA}, output will be \code{NA}}
}
\value{
Length 1 logical vector
}
\description{
Determine whether any values in a vector are equal to zero
}
