% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aou_tables}
\alias{aou_tables}
\title{All of Us tables}
\format{
List of 68 tables. Names refer to All of Us table names. Values are character vector of column names in corresponding table
}
\usage{
aou_tables
}
\description{
Information from allofus package
}
\keyword{datasets}
