% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aou_data.R
\name{aou_workspace_to_bucket}
\alias{aou_workspace_to_bucket}
\title{Move file from workspace to bucket for permanent storage}
\usage{
aou_workspace_to_bucket(
  file,
  directory = FALSE,
  bucket = getOption("aou.default.bucket")
)
}
\arguments{
\item{file}{File in workspace}
}
\value{
File silently exported
}
\description{
Move file from workspace to bucket for permanent storage
}
