% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{as_date}
\alias{as_date}
\title{Convert input to date}
\usage{
as_date(x)
}
\arguments{
\item{x}{Vector containing year, month, day}
}
\value{
Date vector in YYYY-MM-DD format
}
\description{
Convert input to date
}
