% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{as_fct}
\alias{as_fct}
\title{Convert vector to a factor}
\usage{
as_fct(x, levels = NULL)
}
\arguments{
\item{x}{Vector}

\item{levels}{Factor levels. Enter as character vector}
}
\value{
Factor vector
}
\description{
Convert vector to a factor
}
