% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{as_formula}
\alias{as_formula}
\alias{create_formula}
\title{Coerce string to formula}
\usage{
as_formula(x, env = parent.frame())

create_formula(lhs, rhs)
}
\arguments{
\item{x}{String containing formula}

\item{env}{Formula environment. Default is \code{parent.frame()}}

\item{lhs, rhs}{Variables for left and right hand side of formula. Enter as character vectors}
}
\value{
Formula

Formula
}
\description{
Coerce string to formula

Create formula from character vectors for left and right hand side variables
}
