% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{as_ggplot}
\alias{as_ggplot}
\alias{as_ggplot_ggplotify}
\title{Convert plot to ggplot object}
\usage{
as_ggplot(x)

as_ggplot_ggplotify(x, ...)
}
\arguments{
\item{x}{Code to generate base plot}

\item{...}{Arguments passed to \code{ggplotify::as.ggplot}}
}
\value{
ggplot object
}
\description{
Convert plot to ggplot object

Convert input to ggplot object
}
