% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectors.R
\name{as_numeric}
\alias{as_numeric}
\alias{as_integer}
\title{Convert a vector to numeric}
\usage{
as_numeric(x)

as_integer(x)
}
\arguments{
\item{x}{Vector}
}
\value{
Numeric vector with same length as input
}
\description{
Similar to \code{base::as.numeric} but suppresses warnings

Similar to \code{base::as.integer} but suppresses warnings
}
