% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{auc}
\alias{auc}
\title{Area under the curve}
\usage{
auc(
  df,
  formula = NULL,
  x = NULL,
  y = NULL,
  interval = NULL,
  method = c("trapezoidal", "step", "spline", "all")
)
}
\arguments{
\item{df}{Data frame containing dependent variable (\code{y}) and independent variable (\code{x})}

\item{formula}{Formula entered as y ~ x}

\item{x, y}{Column names in \code{df} for x and y axis respectively. Enter as quoted or unquoted variable name}

\item{interval}{Minimum and maximum x values. Range is inclusive (i.e. []). Enter as length 2 numeric vector}

\item{method}{Method to calculate area under the curve. Options: \code{"trapezoidal"} (default), \code{"step"}, \code{"spline"}, \code{"all"}}
}
\value{
Length 1 numeric vector except when \code{method = "all"}, in which case output is a data frame with columns "method" and "auc"
}
\description{
Area under the curve
}
