% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{base_to_ggplot}
\alias{base_to_ggplot}
\title{Convert base plot to ggplot object}
\usage{
base_to_ggplot(expr, envir = parent.frame())
}
\arguments{
\item{expr}{Plotting expression}

\item{envir}{Environment in which \code{expr} should be evaluated. Default is parent frame}
}
\value{
ggplot
}
\description{
Convert base plot to ggplot object
}
