% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_axis.R
\name{breaks_extended}
\alias{breaks_extended}
\title{Calculate break positions}
\usage{
breaks_extended(x, n = 5, expanded = TRUE, nice = c(1, 5, 2, 2.5, 4, 3))
}
\arguments{
\item{x}{Limits or raw data values}

\item{n}{Desired number of axis breaks. Default is \code{5}}

\item{expanded}{If \code{TRUE} (default), first breaks is less than \code{min(x)} and last breaks is greater than \code{max(x)}}

\item{nice}{Numeric vector of "nice" numbers}
}
\value{
Numeric vector of break positions. \code{breaks_extended(x, n)} is equivalent to \code{scales::breaks_extended(n = n, only.loose = TRUE)(x)}
}
\description{
Functionality from \code{extended} function in labelled package
}
