% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_axis.R
\name{breaks_linear}
\alias{breaks_linear}
\title{Create axis breaks for continuous variable}
\usage{
breaks_linear(
  n = 4,
  breaks_fn = pretty,
  min_breaks = 4,
  max_breaks = 6,
  include_negative = FALSE
)
}
\arguments{
\item{n}{Desired number of axis breaks. Enter as length 1 numeric. Default is \code{4}}

\item{breaks_fn}{Function to generate breaks. Default is \code{pretty.} Alternative is \code{breaks_extended}}

\item{min_breaks}{Minimum number of possible axis breaks. Enter as length 1 numeric. Default is \code{4}}

\item{max_breaks}{Maximum number of breaks allowed. Enter as length 1 numeric. Default is \code{6}}

\item{include_negative}{If \code{FALSE} (default), negative numbers are excluded from breaks}
}
\value{
Function. Enter as \code{scale_y_continuous(breaks = breaks_linear())}
}
\description{
Create axis breaks for continuous variable
}
