% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_axis.R
\name{breaks_log}
\alias{breaks_log}
\title{Calculate position of log breaks}
\usage{
breaks_log(base = 10, n = 5, breaks_fn = breaks_extended, log_fn = log)
}
\arguments{
\item{base}{Log base. Enter as length 1 numeric. Default is \code{10}}

\item{n}{Desired number of tick marks. Enter as length 1 numeric. Default is \code{5}}

\item{breaks_fn}{Function to generate breaks. Must accept a numeric vector as input and return a numeric vector as output. Default is \code{breaks_extended}}

\item{log_fn}{Logarithm function. Must have argument named \code{base}. Default is \code{log}}
}
\value{
Function that can be used to generate numeric vector of break positions on log scale
}
\description{
Calculate position of log breaks
}
