% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_coerce.R
\name{cbind_list}
\alias{cbind_list}
\title{Combine list of vectors column-wise to create data frame}
\usage{
cbind_list(x)
}
\arguments{
\item{x}{List of vectors. Each vector must have same length or must share a common multiple (see examples for details). To convert a list of vectors to rows of a data frame, use \code{rbind_list()}}
}
\value{
Data frame with elements of input list (vectors) forming columns of output data frame
}
\description{
Opposite of \code{as.list.data.frame}
}
