% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_descriptive.R
\name{CI}
\alias{CI}
\title{Confidence interval difference from mean}
\usage{
CI(x, ci = 0.95)
}
\arguments{
\item{x}{Numeric vector}

\item{ci}{Type of confidence interval. Enter as length 1 numeric 0-1. Default is \code{0.95}}
}
\value{
Length 1 numeric vector containing the absolute difference between mean and the upper (or lower) bound of CI
}
\description{
CI = z*SE
}
