% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_descriptive.R
\name{CI_prop}
\alias{CI_prop}
\title{Confidence interval for proportion}
\usage{
CI_prop(
  n,
  total,
  ci = 0.95,
  method = c("wilson", "exact", "wald", "wald_corrected", "agresti_coull", "exact",
    "asymptotic", "all")
)
}
\arguments{
\item{n}{Numerator of proportion. Enter as length 1 integer}

\item{total}{Denominator of proportion. Enter as length 1 integer}

\item{ci}{Type of confidence interval. Enter as length 1 numeric 0-1. Default is \code{0.95}}

\item{method}{Method for calculating confidence interval. Options include: \code{"wilson"} (default, used by Prism), \code{"wald"}, \code{"wald_corrected"}, \code{"agresti_coull"}, \code{"exact"}, \code{"asymptotic"}, \code{"all"}}
}
\value{
Length 3 numeric vector containing proportion, lower bound of CI, and upper bound of CI (in that order). If \code{method = "all"}, output is a matrix with 1 row for each method
}
\description{
Functionality from Frank Harrell's excellent package HMisc
}
