% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{clean_names}
\alias{clean_names}
\title{Clean column names for data frame}
\usage{
clean_names(x, prefix = "V", max_width = 100L, unique = TRUE, lower = FALSE)
}
\arguments{
\item{x}{Character vector of column names}

\item{prefix}{Prefix for names that begin with non-letter. Default is \code{"V"}}

\item{max_width}{Maximum number of characters in names. Default is \code{100}}

\item{unique}{If \code{TRUE} (default), names are made unique}

\item{lower}{If \code{TRUE}, output converted to lowercase. Default is \code{FALSE}}
}
\value{
Character vector containing names
}
\description{
Clean column names for data frame
}
