% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{clr_ramp}
\alias{clr_ramp}
\alias{clr_continuous}
\alias{clr_continuous_grey}
\title{General color ramp function}
\usage{
clr_ramp(x = "light", saturation = 1, ...)

clr_continuous(..., n = 20, reverse = FALSE)

clr_continuous_grey(n = 20, start = 0.3, end = 0.8, reverse = FALSE)
}
\arguments{
\item{x}{Colors to generate color ramp. Enter as a character vector of hexadecimal codes (or color names) or 1 of the following length 1 character vectors: \code{"light"} (default), \code{"medium"}, \code{"dark"}, \code{"flowjo"}, \code{"turbo"}, \code{"viridis"}, \code{"cividis"}, \code{"mako"}, \code{"plasma"}, \code{"inferno"}, \code{"rocket"}}

\item{saturation}{Saturation factor. Only relevant when \code{x = "flowjo"}. Enter as length 1 numeric 0-1. Default is \code{1}}

\item{...}{Colors. If 3 colors are entered, output is a continuous scale using 2nd color as midpoint. Can also be one of the following palettes: \code{"light"}, \code{"medium"}, \code{"dark"}, \code{"flowjo"}, \code{"turbo"}, \code{"viridis"}, \code{"cividis"}, \code{"mako"}, \code{"plasma"}, \code{"inferno"}, \code{"rocket"}}

\item{n}{Number of colors. Enter as length 1 numeric. Default is \code{20}}

\item{reverse}{Whether to reverse default color order. If \code{FALSE} (default), colors are ordered from darkest to lightest. If \code{TRUE}, colors are ordered from lightest to darkest}

\item{start, end}{Hues used for first and last colors, respectively. Enter as length 1 numeric, 0-1. Defaults are \code{start = 0.3} and \code{end = 0.8}}
}
\value{
Function with argument \code{n} that expects a length 1 numeric indicating the desired number of colors to generate

Character vector of hexadecimal codes. Length determined by \code{n}
}
\description{
Rewritten version of \code{grDevices::grey.colors}
}
