% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{comorb_lookup}
\alias{comorb_lookup}
\title{Lookup table for comorbidity diagnoses}
\format{
Data frame
}
\usage{
comorb_lookup
}
\description{
Lookup table for comorbidity diagnoses
}
\keyword{datasets}
