% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_cat.R
\name{compare_cat}
\alias{compare_cat}
\title{Compare categorical variable across 2 or more groups}
\usage{
compare_cat(
  df,
  grouping_var,
  split_by = NULL,
  exclude_vars = NULL,
  na_rm = TRUE,
  as_df = TRUE,
  max_n_unique = 20
)
}
\arguments{
\item{df}{Data frame in long format}

\item{grouping_var}{Column in \code{df} that contains groups (i.e., columns in output table). Enter as quoted or unquoted column name}

\item{split_by}{Character vector of column names in \code{df} that will be used to split data to perform separate analysis. No yet incorporated}

\item{exclude_vars}{Character vector of column names in \code{df} to exclude from analyses}

\item{na_rm}{If \code{TRUE} (default), missing values for outcome, grouping, and subgrouping variables are removed from analyses}

\item{as_df}{If \code{TRUE} (default), output will be a data frame. If \code{FALSE}, output will be a list of tables}

\item{max_n_unique}{Maximum number of factor levels for a non-grouping variable to be included in output. Enter as length 1 integer. Default is \code{20}}
}
\value{
Data frame (if \code{as_df = TRUE}) or list of tables (if \code{as_df = FALSE})
}
\description{
Compare categorical variable across 2 or more groups
}
