% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aou_sql_queries.R
\name{concept_code_query}
\alias{concept_code_query}
\alias{condition_query}
\alias{demographics_query}
\alias{lab_concept_query}
\alias{lab_query}
\alias{rx_query}
\alias{rx_with_record_source_query}
\alias{bmi_query}
\alias{death_query}
\alias{survey_query}
\alias{wgs_query}
\alias{icd10_query}
\alias{snomed_query}
\title{Condition concept code query}
\usage{
concept_code_query(
  x,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL,
  dest = "concept_code_query_result.csv"
)

condition_query(
  concept_ids = NULL,
  source_values = NULL,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL,
  dest = "condition_query_result.csv"
)

demographics_query(
  dest = "demographics_query_result.csv",
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)

lab_concept_query(
  x,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL,
  dest = "lab_concept_query_result.csv"
)

lab_query(
  x,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL,
  dest = "lab_query_result.csv"
)

rx_query(
  x,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL,
  dest = "med_query_result.csv"
)

rx_with_record_source_query(
  x,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL,
  dest = "med_with_drug_type_query_result.csv"
)

bmi_query(
  anchor_date_table = NULL,
  before = NULL,
  after = NULL,
  dest = "bmi_query_result.csv"
)

death_query(
  anchor_date_table = NULL,
  before = NULL,
  after = NULL,
  dest = "death_query_result.csv"
)

survey_query(
  x,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL,
  dest = "survey_query.csv"
)

wgs_query(dest = "wgs_query_result.csv")

icd10_query(
  x = NULL,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL,
  dest = "icd10_query_result.csv"
)

snomed_query(
  x = NULL,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL,
  dest = "snomed_query_result.csv"
)
}
\arguments{
\item{x}{Character vector of SNOMED condition codes}

\item{anchor_date_table}{Data frame with 2 columns: person_id, anchor_date. Time window can be defined around anchor_date using \code{before} and \code{after}}

\item{before}{integer >= 0. Dates prior to anchor_date + before will be excluded}

\item{after}{integer >= 0. Dates after anchor_date + after will be excluded}

\item{concept_ids}{Numeric vector with condition concept ids}

\item{source_values}{Character vector with condition source values}
}
\value{
Data frame with columns: person_id, condition_start_date, concept_name, condition_concept_id

Data table with columns: person_id, condition_start_date, condition_concep_id, condition_source_value

Data frame with columns: person_id, date_of_birth, race, ethnicity, sex

Data frame with columns: person_id, measurement_date, value_as_number, value_as_concept

Data frame with columns: person_id, measurement_date, value_as_number

Data frame with columns: person_id, drug_exposure_start_date

Data frame with columns: person_id, drug_exposure_start_date, record_source

Data frame with the following columns: person_id, measurement_date, bmi

Data frame with the following columns: person_id, death_entry_date

Data frame with the following columns: person_id, survey_response, survey_date

Data frame with the following columns: person_id

Data frame with columns: person_id, condition_start_date, condition_source_value

Data frame with columns: person_id, condition_start_date, condition_source_value
}
\description{
Example: concept_code_query(c(441641, 4014295))

Example: condition_query(concept_ids = c(710706, 705076), source_values = "U09.9")

Example: lab_concept_query(c(586520, 586523, 586525))

Example: lab_query(c("Triglyceride \link{Mass/volume} in Serum or Plasma","Triglyceride \link{Mass/volume} in Blood"))

Example: survey_query("1585860")

Example: icd10_query(c("I21","I21.\%"))
}
