% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_transformation.R
\name{convert}
\alias{convert}
\title{Convert units}
\usage{
convert(x, from, to, pt_per_in = 72.27)
}
\arguments{
\item{x}{Numeric vector of quantities to convert. Units of \code{x} specified in \code{from} argument}

\item{from}{Units for \code{x}. Options: \code{"in"}, \code{"cm"}, \code{"mm"}, \code{"pts"} (or \code{"points"}), \code{"lines"}, \code{"linewidth"} (ggplot-specific units. 1 unit is equivalent to 0.75 mm), \code{"npc"}, \code{"C"}, \code{"F"}, \code{"kg"}, \code{"lbs"}. Enter as length 1 character vector}

\item{to}{Desired output units. Same options as \code{from}. Enter as length 1 character vector}

\item{pt_per_in}{Points per inch. Default is \code{72.27} (used by ggplot2). PowerPoint uses 72.009. Enter as length 1 numeric vector}
}
\value{
Numeric vector expressed in units determined by \code{to} argument
}
\description{
Convert units
}
