% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{cor_pairs}
\alias{cor_pairs}
\title{Calculate correlation coefficient and P value for each pair of columns in data frame}
\usage{
cor_pairs(df, ..., method = c("spearman", "pearson", "kendall"), ci = 0.95)
}
\arguments{
\item{df}{Data frame in wide format (correlation will be calculated between columns) or grouped data frame}

\item{...}{Columns in \code{df} to include in correlation analysis. Enter using tidyselect syntax}

\item{method}{Method of comparison. Options: \code{"spearman"} (default), \code{"pearson"}, \code{"kendall"}}

\item{ci}{Confidence interval. Enter as length 1 numeric 0-1. Default is \code{0.95}}
}
\value{
Data frame with columns "col_1", "col_2", "cor", "cor_lower", "cor_upper", "p", "label", "method", "n", "n_nna", "perc_na"
}
\description{
Calculate correlation coefficient and P value for each pair of columns in data frame
}
\examples{
# cor_pairs(covid)
# cor_pairs(dplyr::group_by(covid, severity))

}
