% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_info.R
\name{count_na}
\alias{count_na}
\title{Number of missing values for each column in data frame}
\usage{
count_na(df, ...)
}
\arguments{
\item{df}{Data frame}

\item{...}{Columns in \code{df} for which number of missing values will be determined. Enter using tidyselect syntax}
}
\value{
Data frame with columns for each variable entered as input to \code{...} and 1 row containing number of missing values for that variable
}
\description{
Number of missing values for each column in data frame
}
