% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{Coxph}
\alias{Coxph}
\alias{Survfit}
\alias{Survdiff}
\title{coxph wrapper}
\usage{
Coxph(df, predictor_var, time_var = "time", outcome_var = "death", ...)

Survfit(df, predictor_var, time_var = "time", outcome_var = "death", ...)

Survdiff(
  df,
  predictor_var,
  time_var = "time",
  outcome_var = "death",
  method = "MH",
  ...
)
}
\arguments{
\item{df}{Data frame}

\item{predictor_var}{Variable(s) to use as covariates in model. Enter as character vector}

\item{time_var}{Variable containing follow up time. Enter as length 1 character vector. Default is \code{"time"}}

\item{outcome_var}{Variable containing outcome variable coded as 1 (event) or 0 (no event/censored). Enter as length 1 character vector. Default is \code{"death"}}

\item{...}{Arguments passed to \code{survival::coxph}}

\item{method}{Method for determining P value. Options: \code{"MH"} (default, Mantel-Haenszel test), \code{"PP"} (Peto and Peto modification of Gehan-Wilcoxon test)}
}
\value{
Output from \code{survival::coxph} (list) as well as "outcome_var", "time_var", "predictor_var", "univariate" (logical), "outcome_binary" (logical. \code{TRUE} if outcome is binary and both outcomes occur in all groups)
}
\description{
coxph wrapper

survfit wrapper

survdiff wrapper
}
