% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_classification.R
\name{cutpoint_binary}
\alias{cutpoint_binary}
\title{Generate cutpoints for continuous variable as predictor of binary outcome}
\usage{
cutpoint_binary(df, outcome_var, predictor_var)
}
\arguments{
\item{df}{Data frame}

\item{outcome_var}{Binary outcome variable. Enter as quoted or unquoted variable name}

\item{predictor_var}{Continuous predictor variable. Enter as quoted or unquoted variable name}
}
\value{
Data frame with columns "outcome_var" (outcome variable), "predictor_var" (predictor variable), "cutpoint", "type" (direction of predictor variable used to determine positivity), "sens", "spec", "auc", "n", "above_cp", "perc_above_cp", "below_cp", "perc_below_cp"
}
\description{
Functionality from pROC package
}
