% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{cutpoint_survival}
\alias{cutpoint_survival}
\title{Determine optimal cutpoint for a continuous variable as a predictor of survival}
\usage{
cutpoint_survival(
  df,
  predictor_var,
  time_var = "time",
  outcome_var = "death",
  min_prop = 0.1
)
}
\arguments{
\item{df}{Data frame}

\item{predictor_var}{Continuous predictor variable(s). Enter as character vector}

\item{time_var}{Variable containing follow up time. Enter as length 1 character vector. Default is \code{"time"}}

\item{outcome_var}{Variable containing outcome variable coded as 1 (event) or 0 (no event/censored). Enter as length 1 character vector. Default is \code{"death"}}

\item{min_prop}{Minimum proportion of all subjects in smaller of 2 groups when stratifying by cutpoint. Enter as length 1 numeric 0-1. Default is \code{0.1}}
}
\value{
Data frame with columns "outcome_var", "time_var", "predictor_var", "cutpoint"
}
\description{
Determine optimal cutpoint for a continuous variable as a predictor of survival
}
