% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{day}
\alias{day}
\alias{month}
\alias{year}
\title{Extract day of month from date}
\usage{
day(x)

month(x)

year(x)
}
\arguments{
\item{x}{Vector of dates}
}
\value{
Integer vector with same length as input
}
\description{
Extract day of month from date

Extract month from date

Extract 4 digit year from date
}
\examples{
# day(covid$date_admission)

}
