% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_coerce.R
\name{df_to_matrix}
\alias{df_to_matrix}
\title{Convert data frame to matrix}
\usage{
df_to_matrix(df, rownames = NULL, col_to_rownames = NULL)
}
\arguments{
\item{df}{Data frame}

\item{rownames}{Row names for output matrix. Enter as character vector with length equal to \code{nrow(df)}. If specified, \code{col_to_rownames} must be \code{NULL}}

\item{col_to_rownames}{Column in \code{df} containing row names for output matrix. Note: column containing row names will not be included in output matrix. Enter as length 1 character (column name) or integer (column number) vector. If specified, \code{rownames} must not be specified}
}
\value{
Matrix. If \code{col_to_rownames = NULL}, dimensions of output matrix will be identical to dimensions of \code{df}. If \code{col_to_rownames} is specified, output matrix will have \code{nrow(df)} rows and \code{ncol(df) - 1L} columns
}
\description{
Functionality from \code{base::data.matrix}
}
\examples{
\dontrun{
# df_to_matrix(mtcars)
}

}
