% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_subset.R
\name{Distinct}
\alias{Distinct}
\title{Remove duplicate rows}
\usage{
Distinct(df, ..., keep_all = TRUE)
}
\arguments{
\item{df}{Data frame}

\item{...}{Columns used to detect duplicates. Enter using tidyselect syntax. Default uses all columns}

\item{keep_all}{If \code{TRUE} (default), all columns are included in output. If \code{FALSE}, output limited to columns specified in \code{...}}
}
\value{
Data frame with duplicate rows removed
}
\description{
Identical to \code{dplyr::distinct} but with \code{.keep_all = TRUE} by default
}
