% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\docType{data}
\name{.golden_ratio}
\alias{.golden_ratio}
\title{Golden ratio (phi)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
.golden_ratio
}
\description{
Golden ratio (phi)
}
\keyword{datasets}
