% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programming.R
\name{dots_as_quoted}
\alias{dots_as_quoted}
\alias{dots_as_unquoted}
\title{Collect dots as quoted terms}
\usage{
dots_as_quoted(..., .use_names = FALSE)

dots_as_unquoted(..., .use_names = FALSE)
}
\arguments{
\item{...}{Comma separated list of quoted or unquoted terms}

\item{.use_names}{If \code{TRUE}, output includes any names entered in \code{...}}
}
\value{
Character vector (terms entered in dots are not evaluated or interpreted)

List of unquoted terms. Input to \code{...} is not evaluated or interpreted
}
\description{
Collect dots as quoted terms

Collect dots as unquoted terms
}
