% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_axis.R
\name{draw_axis_clean}
\alias{draw_axis_clean}
\title{Grob to draw clean axis}
\usage{
draw_axis_clean(
  break_positions,
  break_labels,
  axis_position,
  theme,
  angle = NULL
)
}
\arguments{
\item{break_positions}{position of ticks}

\item{break_labels}{labels at ticks}

\item{axis_position}{Axis position. Options: \code{"left"} (default for y axis), \code{"bottom"} (default for x axis), \code{"right"}, \code{"top"}}

\item{theme}{Complete theme}

\item{angle}{Angle for axis tick labels}
}
\value{
Called by \code{guide_gengrob.clean}
}
\description{
Functionality from Charlotte Dawson's excellent package ggprism
}
