% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_info.R
\name{dupe_cols}
\alias{dupe_cols}
\title{Identify columns with identical values}
\usage{
dupe_cols(df, ignore_class = FALSE)
}
\arguments{
\item{df}{Data frame}

\item{ignore_class}{If \code{TRUE}, all columns are converted to characters prior to analysis}
}
\value{
Character vector of column names whose values are duplicated in another column in \code{df}. If no columns are duplicated, output is \code{character(0)}
}
\description{
Identify columns with identical values
}
