% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectors.R
\name{duplicated_incl_first}
\alias{duplicated_incl_first}
\title{Include the first occurrence of the subsequently duplicated value in duplicated output}
\usage{
duplicated_incl_first(x)
}
\value{
Logical vector in which any/all duplicated values are \code{TRUE} and all unique values are \code{FALSE}
}
\description{
Similar to \code{duplicated} but is also \code{TRUE} for first occurence of duplicated value
}
