% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{earliest_sig_surv}
\alias{earliest_sig_surv}
\title{Evaluate all time points to identify time point when curves first separate}
\usage{
earliest_sig_surv(
  df,
  predictor_var,
  time_var = "time",
  outcome_var = "death",
  time_by = 1L
)
}
\arguments{
\item{df}{Data frame or grouped data frame}

\item{predictor_var}{Variable(s) to use as covariates in model. Enter as character vector}

\item{time_var}{Name of variable in \code{df} containing follow up time as numeric. Enter as length 1 character vector. Default is \code{"time"}}

\item{outcome_var}{Name of variable in \code{df} containing outcome coded as 1 (event) or 0 (no event/censored). Enter as length 1 character vector. Default is \code{"death"}}

\item{time_by}{Interval between time points. Enter as length 1 integer vector. Default is \code{1L}}
}
\value{
Data frame with columns "time_var, "time", "variable", "predictor", "hr", "hr_lower", "hr_upper", "p_lrt", "p_wald", "n_events", "n", "n_events_ever", "cfd_events" (proportion of events that occurred by a given time as as a proportion of total events that ever occurred)
}
\description{
Evaluate all time points to identify time point when curves first separate
}
