% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_errorbar.R
\name{errorbar}
\alias{errorbar}
\title{Error bars}
\usage{
errorbar(
  summary_fn = Mean,
  error_fn = SE,
  error_limits = "upper",
  line_thickness = 1,
  width = 0.1,
  dodge = NULL,
  ...
)
}
\arguments{
\item{summary_fn}{Function used to generated summary estimate for y. Default is \code{Mean}}

\item{error_fn}{Function used to determine error limits for y. Default is \code{SE}}

\item{error_limits}{Options: \code{"upper"} (default), \code{"lower"}, \code{"both"}}

\item{line_thickness}{Thickness of error bar lines in pts units. Enter as length 1 numeric. Default is \code{1}}

\item{width}{Width of error bar. Enter as length 1 numeric. Default is \code{0.1}}

\item{dodge}{Dodge width. Enter as length 1 numeric. Default is \code{0.9}}

\item{...}{Arguments passed to \code{stat_summary}}
}
\value{
ggproto object. Add error bars to plot using \code{plot + errorbar()}
}
\description{
Error bars
}
