% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_values.R
\name{format_p_value}
\alias{format_p_value}
\title{Format P values}
\usage{
format_p_value(
  p,
  prefix = "P",
  trim_ws = FALSE,
  show_leading_0 = TRUE,
  min_digits = 2
)
}
\arguments{
\item{p}{Numeric vector containing P values}

\item{prefix}{Prefix for P value. Default is \code{"P"}}

\item{trim_ws}{If \code{FALSE} (default), white space is allowed in output. If \code{TRUE}, white space is removed from output}

\item{show_leading_0}{If \code{TRUE} (default), 0 before the decimal is included in output. If \code{FALSE}, 0 before decimal is removed in output (for example, .01)}

\item{min_digits}{Number of digits included after decimal when P > 0.06. Enter as length 1 numeric. Default is \code{2}}
}
\value{
Character vector with same length as input
}
\description{
Format P values
}
