% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_errorbar.R
\docType{data}
\name{GeomErrorUpper}
\alias{GeomErrorUpper}
\title{ggproto for \code{geom_errorupper}}
\format{
An object of class \code{GeomErrorUpper} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 8.
}
\usage{
GeomErrorUpper
}
\value{
ggproto object
}
\description{
ggproto for \code{geom_errorupper}
}
\keyword{datasets}
