% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sig_geoms.R
\docType{data}
\name{GeomSig}
\alias{GeomSig}
\title{ggproto for Geom to create significance annotation}
\format{
An object of class \code{GeomSig} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 8.
}
\usage{
GeomSig
}
\description{
ggproto for Geom to create significance annotation
}
\keyword{datasets}
