% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_info.R
\name{get_dupes}
\alias{get_dupes}
\alias{get_dupes.default}
\alias{get_dupes.data.frame}
\alias{get_dupes.grouped_df}
\title{Identify duplicate values in a vector or rows in a data frame}
\usage{
get_dupes(x, ...)

\method{get_dupes}{default}(x, ...)

\method{get_dupes}{data.frame}(x, ...)

\method{get_dupes}{grouped_df}(x, ...)
}
\arguments{
\item{x}{Data frame or vector}

\item{...}{Columns to search for duplicates. Enter using tidyselect syntax. Default uses all columns. Only relevant when \code{x} is a data frame}
}
\value{
Data frame with columns "duplicated_value", "n_dupes", "positions"
}
\description{
Searches for duplicate rows
}
