% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_extract.R, R/plot_info.R
\name{get_plot_aes}
\alias{get_plot_aes}
\alias{get_plot_vars}
\title{Extract \code{aes} arguments for each plot layer}
\usage{
get_plot_aes(x)

get_plot_vars(x)
}
\arguments{
\item{x}{ggplot object}
}
\value{
List of aes/variables for each plot layer. Output length equal to number of plots layers
}
\description{
Extract \code{aes} arguments for each plot layer

Extract plot variables
}
