% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_info.R
\name{get_plot_limits}
\alias{get_plot_limits}
\title{Overall limits for plot (data + axis)}
\usage{
get_plot_limits(x, by_facet = FALSE)
}
\arguments{
\item{by_facet}{If \code{TRUE} and \code{x} contains facets, output will be a data frame with 1 row per facet and columns "panel", "row", "col", "vars" (facet variables), "xmin", "xmax", "ymin", "ymax". If \code{FALSE} (default), output is a list containing "x" and "y", each a length 2 numeric vector representing the x and y axis limits across all facets. Only relevant when \code{x} contains facets}
}
\value{
If \code{x} doesn't contain facets and/or \code{by_facet = FALSE}, will be a list containing \code{x} and \code{y} (each a length 2 numeric vector containing min and max across all facets). If plot has multiple facets and \code{by_facet = TRUE}, output will be a data frame with 1 row per facet and columns "panel", "row", "col", "vars" (facet variables), "xmin", "xmax", "ymin", "ymax"
}
\description{
Overall limits for plot (data + axis)
}
