% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_extract.R
\name{get_plot_axis_trans}
\alias{get_plot_axis_trans}
\title{Extract transformation/inverse functions applied to x and y axes}
\usage{
get_plot_axis_trans(x)
}
\arguments{
\item{x}{ggplot object}
}
\value{
List of length 2 ("x", "y"). "x" and "y" each consist of a list of length 2 ("trans", "inverse")
}
\description{
Extract transformation/inverse functions applied to x and y axes
}
