% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_info.R
\name{get_plot_axis_vars}
\alias{get_plot_axis_vars}
\alias{get_plot_axis_labels}
\alias{get_plot_axis_breaks}
\title{Get variables used to generate x and y axes}
\usage{
get_plot_axis_vars(x, axis = "both")

get_plot_axis_labels(x, axis = "both")

get_plot_axis_breaks(x, axis = "both", units = "data")
}
\arguments{
\item{x}{ggplot object}

\item{axis}{Options: \code{"both"} (default), \code{"x"}, \code{"y"}}

\item{units}{Options: \code{"data"} (default, raw data units), \code{"npc"}}
}
\value{
Raw data input variables used to generate x and/or y axis

Character vector containing axis tick labels
}
\description{
Get variables used to generate x and y axes

Get axis tick labels

Get axis breaks
}
