% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_extract.R
\name{get_plot_color_scale}
\alias{get_plot_color_scale}
\title{Extract color scale}
\usage{
get_plot_color_scale(x)
}
\arguments{
\item{x}{ggplot object}
}
\value{
Same class as scale_color output
}
\description{
Extract color scale
}
