% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_info.R
\name{get_plot_data_limits}
\alias{get_plot_data_limits}
\alias{get_plot_axis_limits}
\title{Data limits (in raw data units) for all data frames used to generate plot}
\usage{
get_plot_data_limits(x, axis = "both")

get_plot_axis_limits(x, axis = "both", units = "data")
}
\arguments{
\item{x}{ggplot object}

\item{axis}{Options: \code{"both"} (default), \code{"x"}, \code{"y"}}

\item{units}{Units for output. Options: \code{"data"} (default, raw data), \code{"npc"}}
}
\value{
List containing \code{x} and \code{y}, each a length 2 numeric vector with limits in raw data units

If \code{axis = "x"} or \code{axis = "y"}, output is a length 2 numeric vector containing limits. If \code{axis = "both"}, output is a list containing \code{x} and \code{y}, each a length 2 numeric vector containing axis limits
}
\description{
Limits for axis line. When clip = "off", data limits can extend past axis limits
}
