% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_axis.R
\name{ggplot_add.axis_clean}
\alias{ggplot_add.axis_clean}
\title{Add continuous axis scale to ggplot}
\usage{
\method{ggplot_add}{axis_clean}(object, plot, object_name)
}
\arguments{
\item{object}{Output from \code{scale_continuous}}

\item{plot}{ggplot object}

\item{object_name}{Name of object}
}
\value{
ggplot object. Needed for \code{scale_continuous}. Not called directly by user. Must be exported
}
\description{
Add continuous axis scale to ggplot
}
