% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_axis.R
\name{ggplot_add.axis_x10}
\alias{ggplot_add.axis_x10}
\title{Add axis scale that removes " x 10^x" from axis labels}
\usage{
\method{ggplot_add}{axis_x10}(object, plot, object_name)
}
\arguments{
\item{object}{New scale to add}

\item{plot}{ggplot object}

\item{object_name}{Name of object}
}
\value{
ggplot object. Needed for \code{scale_x10}. Not called directly by user. Must be exported
}
\description{
Add axis scale that removes " x 10^x" from axis labels
}
