% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_manipulation.R
\name{ggplot_add.axis_x_rotated}
\alias{ggplot_add.axis_x_rotated}
\title{Rotate x axis labels}
\usage{
\method{ggplot_add}{axis_x_rotated}(object, plot, object_name)
}
\arguments{
\item{object}{Output from \code{scale_continuous}}

\item{plot}{ggplot object}

\item{object_name}{Name of object}
}
\value{
Not called directly by user
}
\description{
Rotate x axis labels
}
