% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{ggplot_add.new_color}
\alias{ggplot_add.new_color}
\title{Add new color scale}
\usage{
\method{ggplot_add}{new_color}(object, plot, object_name)
}
\arguments{
\item{object}{New scale to add}

\item{plot}{ggplot object}

\item{object_name}{Name of object}
}
\value{
Not called directly by user. Must be exported. Needed for \code{scale_fill_new} and \code{scale_color_new}
}
\description{
Add new color scale
}
