% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{Glm}
\alias{Glm}
\title{Logistic regression}
\usage{
Glm(
  df,
  formula = NULL,
  y = NULL,
  x = NULL,
  family = stats::binomial,
  na.rm = TRUE,
  ...
)
}
\arguments{
\item{df}{Data frame}

\item{formula}{Enter in outcome ~ predictor (y ~ x) format. Set \code{formula = NULL} (default) if using \code{x} and \code{y}}

\item{y}{Outcome variable Enter as character vector. Length must be 1. Ignored if \code{formula} is not \code{NULL}}

\item{x}{Predictor variable(s). Length can be > 1 (multi-variate). Ignored if \code{formula} is not \code{NULL}}

\item{family}{Function passed to family argument of glm. Default is \code{stats::binomial}}

\item{na.rm}{If \code{TRUE} (default), \code{NA} will be removed}

\item{...}{Not used}
}
\value{
glm object
}
\description{
Modified version of \code{stats::glm}
}
