% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{glm_tidy}
\alias{glm_tidy}
\title{Tidy glm output}
\usage{
glm_tidy(x, ci = 0.95)
}
\arguments{
\item{x}{glm object}

\item{ci}{Confidence interval. Enter as length 1 numeric 0-1. Default is \code{0.95}}
}
\value{
Data frame with columns "outcome", "variable", "predictor", "or", "or_lower", "or_upper", "p", "type" (Univariate or Multivariate), "method" (glm), "covariates"
}
\description{
Tidy glm output
}
