% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_contains}
\alias{str_contains}
\title{Detect pattern in string}
\usage{
str_contains(x, pattern, ignore_case = FALSE, invert = FALSE, ...)
}
\arguments{
\item{ignore_case}{If \code{FALSE} (default), search is case sensitive (i.e., output is only \code{TRUE} if the exact input to \code{pattern} is occurs is found in `x)}

\item{...}{Arguments passed to \code{grepl}}
}
\description{
Detect pattern in string
}
