% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_axis.R
\name{guide_gengrob.clean}
\alias{guide_gengrob.clean}
\title{Method for for \code{guide_gengrob} applied to "clean" object}
\usage{
\method{guide_gengrob}{clean}(guide, theme)
}
\arguments{
\item{guide}{Guide}

\item{theme}{Theme}
}
\value{
Not called directly. Used in \code{guide_clean_axis} function. Must be exported
}
\description{
Method for for \code{guide_gengrob} applied to "clean" object
}
