% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{hr_ci}
\alias{hr_ci}
\alias{hr_ci.default}
\alias{hr_ci.coxph}
\alias{hr_ci.data.frame}
\alias{hr_ci.grouped_df}
\title{Hazard ratio, confidence interval, P values}
\usage{
hr_ci(
  x,
  predictor_var = NULL,
  time_var = "time",
  outcome_var = "death",
  ci = 0.95,
  ...
)

\method{hr_ci}{default}(
  x,
  predictor_var = NULL,
  time_var = "time",
  outcome_var = "death",
  ci = 0.95,
  ...
)

\method{hr_ci}{coxph}(
  x,
  predictor_var = NULL,
  time_var = "time",
  outcome_var = "death",
  ci = 0.95,
  ...
)

\method{hr_ci}{data.frame}(
  x,
  predictor_var = NULL,
  time_var = "time",
  outcome_var = "death",
  ci = 0.95,
  ...
)

\method{hr_ci}{grouped_df}(
  x,
  predictor_var = NULL,
  time_var = "time",
  outcome_var = "death",
  ci = 0.95,
  ...
)
}
\arguments{
\item{x}{Data frame, grouped data frame, coxph}

\item{predictor_var}{Name of column(s) in \code{x} that contains covariates in model. Enter as character vector}

\item{time_var}{Name of column in \code{x} that contains follow up time. Enter as length 1 character vector. Default is \code{"time"}}

\item{outcome_var}{Name of column in \code{x} that contains outcome variable coded as 1 (event) or 0 (no event/censored). Enter as length 1 character vector. Default is \code{"death"}}

\item{ci}{Confidence interval. Enter as length 1 numeric 0-1. Default is \code{0.95}}

\item{...}{Arguments passed to class-specific \code{hr_ci} function}
}
\value{
Data frame with columns "predictor_var", "predictor", "hr", "hr_lower", "hr_upper", "p_lrt", "p_wald", "n", "n_events"
}
\description{
Hazard ratio, confidence interval, P values

\code{hr_ci} - default method

\code{hr_ci} - coxph

\code{hr_ci} - data frame

\code{hr_ci} - grouped_dr
}
