% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Lib}
\alias{Lib}
\title{Load packages from cran}
\usage{
Lib(
  ...,
  lib.loc = NULL,
  warn.conflicts,
  mask.ok,
  exclude,
  include.only,
  attach.required = missing(include.only)
)
}
\description{
Load packages from cran
}
