% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{interleave}
\alias{interleave}
\title{Interleave vectors or lists}
\usage{
interleave(..., coerce_fn = as.character)
}
\arguments{
\item{...}{Either a comma separated list of vectors not wrapped in \code{c()} or a list containing vectors}

\item{coerce_fn}{Function applied to each input vector prior to interleaving. Default is \code{as.character}}
}
\value{
Vector or list
}
\description{
Interleave vectors or lists
}
