% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectors.R
\name{is_integerish}
\alias{is_integerish}
\alias{is_continuous}
\title{Determine whether numeric values are nearly equivalent to integers}
\usage{
is_integerish(x)

is_continuous(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Length 1 logical vector. \code{TRUE} if all values in \code{x} are integers. \code{FALSE} if any values in \code{x} are not integers

Length 1 logical vector. \code{TRUE} if all values in \code{x} are non-integer numbers
}
\description{
Determine whether numeric values are nearly equivalent to integers

Determine whether a numeric vector is continuous
}
