% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{is_named}
\alias{is_named}
\title{Determine whether each elements is named}
\usage{
is_named(x)
}
\arguments{
\item{x}{List or vector}
}
\value{
Logical vector with same length as input
}
\description{
Determine whether each elements is named
}
