% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aou_sql_queries.R
\name{lab_query_extended}
\alias{lab_query_extended}
\title{Lab query with extended columns}
\usage{
lab_query_extended(
  x,
  ext_cols = NULL,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL,
  dest = "lab_query_result.csv"
)
}
\arguments{
\item{x}{Character vector or string containing the labs to query}

\item{ext_cols}{Extra columns to return on top of the date and value of the lab}
}
\value{
data.table containing the following columns: person_id, measurement_date, value_as_number, ext_cols
}
\description{
Example: lab_query(c("Triglyceride \link{Mass/volume} in Serum or Plasma","Triglyceride \link{Mass/volume} in Blood"), ext_cols = c("unit_source_value"))
}
