% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_join.R
\name{left_join_safe}
\alias{left_join_safe}
\alias{lj}
\alias{right_join_safe}
\alias{semi_join_safe}
\alias{inner_join_safe}
\alias{anti_join_safe}
\title{Left join with information about number of rows before and after join}
\usage{
left_join_safe(df1, df2, by = NULL, ...)

lj(df1, df2, by = NULL, ...)

right_join_safe(df1, df2, by = NULL, ...)

semi_join_safe(df1, df2, by = NULL, ...)

inner_join_safe(df1, df2, by = NULL, ...)

anti_join_safe(df1, df2, by = NULL, ...)
}
\arguments{
\item{df1, df2}{Primary and secondary data frames}

\item{by}{Key variables shared between \code{df1} and \code{df2}}

\item{...}{Arguments passed to \code{dplyr::anti_join()}}
}
\value{
Merged data frame with message containing number of rows in \code{df1} and output data frame

Merged data frame with message containing number of rows in \code{df2} and output data frame

Merged data frame with message containing number of rows in input and output data frames

Merged data frame with message containing number of rows in input and output data frames

Merged data frame with message containing number of rows in input and output data frames
}
\description{
Left join with information about number of rows before and after join

Alias for left_join_safe

Right join with information about number of rows before and after join

Semi join with information about number of rows before and after join

Inner join with information about number of rows before and after join

Anti join with information about number of rows before and after join
}
