% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{lm_reg_line_equation}
\alias{lm_reg_line_equation}
\title{Create regression formula}
\usage{
lm_reg_line_equation(
  x,
  digits = 2,
  bracket_type = "{",
  mult_sign = "×",
  subtract_sign = "–"
)
}
\arguments{
\item{x}{lm object}

\item{digits}{Number of digits after decimal to include in formula. Default is \code{2}}

\item{bracket_type}{Type of bracket to surround variables in model. Default uses brace}

\item{mult_sign}{String used to indicate multiplication in output. Default is \code{"×"}}

\item{subtract_sign}{String used to indicate multiplication in output. Default is \code{"–"}}
}
\value{
Length 1 character vector
}
\description{
Create regression formula
}
