% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{lm_tidy}
\alias{lm_tidy}
\alias{lm_glance}
\title{Tidy lm output}
\usage{
lm_tidy(x, ci = 0.95)

lm_glance(x)
}
\arguments{
\item{x}{lm object}

\item{ci}{Confidence interval. Enter as length 1 numeric 0-1. Default is \code{0.95}}
}
\value{
Data frame with columns "variable", "predictor", "or", "or_lower", "or_upper", "p"

Data frame with columns "n", "r_sq", "r_sq_adj", "p", "aic", "bic", "univariate", "covariates"
}
\description{
Tidy lm output

Glance lm output
}
