% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_transformation.R
\name{log10_1}
\alias{log10_1}
\alias{lognp}
\title{log10 that leaves 0 unchanged}
\usage{
log10_1(x)

lognp(x, n = abs(Min(c(Min(x), 0))) + 1, base = exp(1))
}
\arguments{
\item{x}{Numeric vector}

\item{n}{Constant added to \code{x} prior to log transformation. Default uses 1 (if \code{x} doesn't contain negative values) or shifts values so that the most negative value is 0, then adds 1 (if \code{x} contains negative values). Enter as length 1 numeric}

\item{base}{Logarithm base. Default is \code{exp(1)}. Enter as length 1 numeric}
}
\value{
Numeric vector with same length as input. Output is \code{NaN} when \code{x < 0}, \code{0} when \code{x == 0}, \code{NA_real_} when \code{is.na(x)}, and \code{log10(x)} for all other values of \code{x}

Numeric vector with same length as input
}
\description{
Similar to \code{base::log1p} but can control amount added to initial raw values (in contrast to log1p which uses 1)
}
