% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{mgsub}
\alias{mgsub}
\title{Substitute multiple pattern/replacement pairs in a string}
\usage{
mgsub(x, pattern, replacement, ..., .sub_fn = gsub)
}
\arguments{
\item{x}{Character vector}

\item{pattern}{Character vector of search patterns to replace in \code{x}}

\item{replacement}{Character vector containing replacement patterns. Will be recycled to \code{length(pattern)}}

\item{...}{Arguments passed to substitution function (either \code{sub} or \code{gsub})}

\item{.sub_fn}{String substitution function. Default is \code{gsub}}
}
\description{
Substitute multiple pattern/replacement pairs in a string
}
