% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_descriptive.R
\name{N}
\alias{N}
\alias{N_na}
\alias{Perc}
\alias{Perc_na}
\alias{Sum}
\alias{Median}
\alias{Mean}
\alias{Min}
\alias{Max}
\alias{Var}
\alias{SD}
\alias{SE}
\alias{CV}
\title{Number of non-missing values}
\usage{
N(x)

N_na(x)

Perc(x)

Perc_na(x)

Sum(x)

Median(x)

Mean(x)

Min(x)

Max(x)

Var(x)

SD(x)

SE(x)

CV(x)
}
\arguments{
\item{x}{Vector}
}
\value{
Length 1 numeric vector
}
\description{
Functionality from \code{stats::median}
}
