% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectors.R
\name{n_per_group}
\alias{n_per_group}
\alias{npg}
\title{Counts for each unique value}
\usage{
n_per_group(x, ...)

npg(x, ...)
}
\arguments{
\item{x}{Vector or data frame}

\item{...}{Columns in \code{x}. Enter using tidyselect syntax. Ignored if \code{x} is a vector}
}
\value{
If \code{x} is a vector, out is a named integer vector with values containing number of observations per level and name representing the corresponding level of \code{x}, listed in descending order of group size with NAs listed last. If \code{x} is a data frame, output is a data frame containing columns "variable", "group" (unique value for variable), "Freq" (number of rows in \code{x} in which variable = group). Similar to output from \code{dplyr::count}
}
\description{
Counts for each unique value

Alias for \code{n_per_group}
}
