% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectors.R
\name{n_unique}
\alias{n_unique}
\title{Determine number of unique values for vector}
\usage{
n_unique(x, na.rm = TRUE, drop = FALSE)
}
\arguments{
\item{x}{Vector}

\item{na.rm}{If \code{TRUE} (default), missing levels are removed}

\item{drop}{If \code{TRUE} and \code{x} is a factor, unused levels are dropped before determining number of levels. Default is \code{FALSE}}
}
\value{
Length 1 integer vector
}
\description{
Determine number of unique values for vector
}
