% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectors.R
\name{na_rm}
\alias{na_rm}
\title{Remove missing values from vector or rows with missing values for certain columns}
\usage{
na_rm(x, cols = names(x))
}
\arguments{
\item{x}{Vector or data frame}

\item{cols}{Columns to search for missing values. Enter as character vector. Default uses all columns in \code{x}. Only relevant when \code{x} is a data frame}
}
\value{
Same class as input
}
\description{
Remove missing values from vector or rows with missing values for certain columns
}
