% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_coerce.R
\name{named_vector_to_df}
\alias{named_vector_to_df}
\title{Convert named vector to data frame with 2 columns (1 column for values, 1 column for named)}
\usage{
named_vector_to_df(x, colname_names = "names", colname_values = "values")
}
\arguments{
\item{x}{Named vector}

\item{colname_names, colname_values}{Column names for names and values, respectively. Enter as length 1 character vector}
}
\value{
Data frame with 2 columns
}
\description{
Convert named vector to data frame with 2 columns (1 column for values, 1 column for named)
}
