% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{names_which}
\alias{names_which}
\title{Get subset of names which contain a specified pattern}
\usage{
names_which(x, pattern, ignore_case = FALSE, ...)
}
\arguments{
\item{x}{Data frame, matrix, list, or named vector}

\item{pattern}{Pattern in \code{names(x)} to search}

\item{ignore_case}{If \code{FALSE} (default), case of both \code{names(x)} and \code{pattern} are ignored in search. If \code{TRUE}, case used in \code{names(x)} and \code{pattern} must match}

\item{...}{Arguments passed to \code{str_contains}}
}
\value{
Character vector containing names that contain \code{pattern}. If no matches found, output has length 0
}
\description{
Get subset of names which contain a specified pattern
}
