% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_info.R
\name{Nrow}
\alias{Nrow}
\alias{seq_nrow}
\title{Faster version of \code{base::nrow}}
\usage{
Nrow(x)

seq_nrow(x)
}
\arguments{
\item{x}{Data frame (doesn't work on matrices)}
}
\value{
Length 1 integer vector
}
\description{
Faster version of \code{base::nrow}

Row indices
}
