% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_values.R
\name{p_adjust}
\alias{p_adjust}
\title{Adjusted P value}
\usage{
p_adjust(x, method = "holm")
}
\arguments{
\item{x}{Numeric vector of unadjusted P values}

\item{method}{Method for P value correction. Options: \code{"holm"} (default), \code{"BH"}, \code{"BY"}, \code{"hochberg"}, \code{"hommel"}, \code{"bonferroni"}, \code{"fdr"} (same as \verb{"BH")}, \code{"none"} (no adjustment). Enter as length 1 character vector. Case doesn't matter}
}
\value{
Numeric vector of adjusted P values with the same length as input
}
\description{
Modified version of \code{stats::p.adjust}
}
