% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_means.R
\name{p_anova}
\alias{p_anova}
\alias{p_kruskal}
\alias{p_tukey}
\title{One-way ANOVA with or without Welch's correction}
\usage{
p_anova(
  df,
  formula = NULL,
  y = NULL,
  x = NULL,
  na.rm = TRUE,
  welch = NULL,
  variance_test = p_levene,
  otherwise = NA_real_,
  ...
)

p_kruskal(
  df,
  formula = NULL,
  y = NULL,
  x = NULL,
  na.rm = TRUE,
  otherwise = NA_real_,
  ...
)

p_tukey(
  df,
  formula = NULL,
  y = NULL,
  x = NULL,
  na.rm = TRUE,
  variance_test = p_F_test,
  ...
)
}
\arguments{
\item{df}{Data frame}

\item{formula}{Enter as y ~ x (continuous variable ~ grouping variable)}

\item{x, y}{Columns containing continuous values for each group. Enter as length 1 character vectors}

\item{na.rm}{If \code{TRUE} (default), missing values for grouping variable are not considered as a distinct group}

\item{welch}{If \code{NULL} (default), variance testing is used to determine whether Welch's correction should be applied. If \code{TRUE}, Welch's correction is performed. If \code{FALSE}, regular ANOVA performed}

\item{variance_test}{Test used to determine whether variance differs by group. Default is \code{p_levene}}

\item{otherwise}{Output if ANOVA fails. Default is \code{NA_real_}}

\item{...}{Not used}
}
\value{
Length 1 numeric vector containing P value

Data frame with comparisons similar to output from \code{compare_means} and \code{p_dunn}
}
\description{
Functionality from \code{stats::oneway.test}
}
