% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{p_cor}
\alias{p_cor}
\alias{cor_test}
\title{P value for correlation test}
\usage{
p_cor(
  df = NULL,
  formula = NULL,
  x = NULL,
  y = NULL,
  method = c("spearman", "pearson", "kendall"),
  ci = 0.95,
  hypothesis_type = c("two.sided", "less", "greater"),
  otherwise = NA_real_
)

cor_test(
  df = NULL,
  formula = NULL,
  x = NULL,
  y = NULL,
  method = c("spearman", "pearson", "kendall"),
  ci = 0.95,
  hypothesis_type = c("two.sided", "less", "greater")
)
}
\arguments{
\item{df}{Data frame}

\item{formula}{Formula in format y ~ x (same as x ~ y)}

\item{x, y}{Column names containing values to correlate. Enter as length 1 character vectors. Can also enter numeric inputs for \code{x} and \code{y} (in which case, \code{df} must be \code{NULL})}

\item{method}{Method for correlation. Options: \code{"spearman"} (default), \code{"pearson"}, \code{"kendall"}}

\item{ci}{Confidence interval. Enter as length 1 numeric 0-1. Default is \code{0.95}}

\item{hypothesis_type}{Type of hypothesis testing to perform. Options: \code{"two.sided"} (default), \code{"less"}, \code{"greater"}. Enter as length 1 character vector}

\item{otherwise}{Value to return if unable to perform correlation test}
}
\value{
Length 1 numeric vector containing P value

List of 4 numeric vectors (each length 1): "cor_coeff", "cor_lower", "cor_upper", "p"
}
\description{
P value for correlation test

Calculate correlation coefficient and 95\% CI
}
