% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_variance.R
\name{p_F_test}
\alias{p_F_test}
\title{F test (aka var.test)}
\usage{
p_F_test(
  df,
  formula = NULL,
  y = NULL,
  x = NULL,
  hypothesis_type = "two.sided",
  otherwise = 0
)
}
\arguments{
\item{df}{Data frame}

\item{formula}{Formula entered as continuous variable ~ grouping variable (y ~ x)}

\item{y}{Continuous variable. Enter as length 1 character vector}

\item{x}{Categorical variable. Enter as length 1 character vector}

\item{hypothesis_type}{Type of hypothesis test to perform. Options: \code{"two.sided"} (default), \code{"less"}, \code{"greater"}. Enter as quoted hypothesis test type}

\item{otherwise}{Output to return if F test fails. Default is \code{0}}
}
\value{
Length 1 numeric vector containing P value
}
\description{
F test (aka var.test)
}
