% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_prop.R
\name{p_mcnemar}
\alias{p_mcnemar}
\title{McNemar's Chi-squared test}
\usage{
p_mcnemar(x, y, corrected = TRUE, otherwise = NA_real_)
}
\arguments{
\item{x, y}{Integer vector for each group}

\item{corrected}{If \code{TRUE} (default), continuity correction applied}

\item{otherwise}{Value returned if P value cannot be calculated. Default is \code{NA_real_}}
}
\value{
Length 1 numeric vector containing P value
}
\description{
Functionality from \code{stats::mcnemar.test}
}
