% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_prop.R
\name{p_prop}
\alias{p_prop}
\alias{p_chi}
\alias{p_fisher}
\title{P value from either chi-squared or Fisher's exact test depending on group size}
\usage{
p_prop(x, test = NULL, n_min_chisq = 5, otherwise = NA_real_, ...)

p_chi(x, yates = FALSE, otherwise = NA_real_, ...)

p_fisher(x, otherwise = NA_real_, ...)
}
\arguments{
\item{x}{Matrix or table containing counts. If \code{x} is a vector of counts, enter as tn, fn, fp, tp. Values entered into matrix by column. Middle 2 numbers must share true/false designation (i.e. either tp and tn or fp and fn)}

\item{test}{Function for testing. Options: NULL (function determined by group size), \code{p_chi}, \code{p_fisher}}

\item{n_min_chisq}{Minimum group size that will be allowed for chi-squared test. Default is \code{5}. Only relevant when test is \code{NULL}}

\item{otherwise}{Value to return if P value can't be determined}

\item{...}{Arguments passed to \code{p_chi()} or \code{p_fisher()}}

\item{yates}{If \code{TRUE} and \code{x} is 2 x 2, Yates' continuity correction is applied}
}
\value{
Length 1 numeric vector containing P value
}
\description{
Functionality from \code{stats::chisq.test}
}
