% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{p_prop_hazards}
\alias{p_prop_hazards}
\title{Test Cox model for proportional hazards}
\usage{
p_prop_hazards(df, predictor_var, time_var = "time", outcome_var = "death")
}
\arguments{
\item{df}{Data frame or coxph object}

\item{predictor_var}{Variable(s) to use as covariates in model. Enter as character vector}

\item{time_var}{Variable containing follow up time. Enter as length 1 character vector. Default is \code{"time"}}

\item{outcome_var}{Variable containing outcome variable coded as 1 (event) or 0 (no event/censored). Enter as length 1 character vector. Default is \code{"death"}}
}
\value{
Data frame with columns "predictor_var" (1 row for global model) and "p". If P < 0.05, proportional hazards assumption is violated
}
\description{
Test Cox model for proportional hazards
}
