% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{pad_leading_zero}
\alias{pad_leading_zero}
\title{Pad numbers with leading zeros}
\usage{
pad_leading_zero(x, n = NULL)
}
\arguments{
\item{x}{Integer or character vector that is coercible to an integer vector}

\item{n}{Number of digits to add. If \code{NULL} (default), determined by longest element in \code{x}}
}
\value{
Character vector with same length as input
}
\description{
Pad numbers with leading zeros
}
