% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar_point.R
\name{plot_bar_point}
\alias{plot_bar_point}
\title{Bar plot with overlying points}
\usage{
plot_bar_point(
  df,
  formula = NULL,
  grouping_var = NULL,
  grouping_var_order = NULL,
  rev_grouping_var_order = FALSE,
  x_order = NULL,
  rev_x_order = FALSE,
  x = NULL,
  y = NULL,
  beeswarm = FALSE,
  beeswarm_method = "smiley",
  width = 0.2,
  beeswarm_width = width,
  n_bins = NULL,
  band_width = 1 + n_bins/5,
  dodge = NULL,
  jitter_width = width,
  colors = c("#0072B5", "#BC3C29", "#999999", "#333333", "#CC79A7", "#56B4E9", "#E69F00",
    "#20854E", "#F0E442", "#7D5FA7", "#97572B", "#005E7A", "#00A0B0", "#8D9D58",
    "#DA3978", "#61D04F", "#FBA27D", "#FFD500", "#00C5CD", "#F7F7F7"),
  color_var = NULL,
  point_shapes = c("circle", "circle"),
  point_size = 5,
  show_points = TRUE,
  small_points = point_size == 0,
  point_alpha = 0.9,
  bar_alpha = 0,
  bar_border_thickness = 0.75,
  point_border_thickness = 1,
  bar_border_color = "black",
  point_border_color = "black",
  bar_width = NULL,
  point_shape_var = NULL,
  show_errorbar = TRUE,
  summary_fn = Mean,
  error_fn = SE,
  errorbar_width_multiplier = 0.5,
  show_legend = FALSE,
  show_legend_points = show_legend,
  show_legend_bars = show_legend,
  legend_title = "",
  y_scale = "regular",
  y_axis_breaks = NULL,
  y_axis_labels = NULL,
  x_axis_breaks = NULL,
  x_axis_labels = NULL,
  y_title = waiver(),
  y_axis_title = y_title,
  x_title = waiver(),
  x_axis_title = x_title,
  plot_title = NULL,
  show_sig = TRUE,
  stars = TRUE,
  sig_font_color = "black",
  sig_bar_color = "black",
  sig_bar_thickness = 0.6,
  show_ns = TRUE,
  ns_symbol = "ns",
  sig_star_size = 20,
  sig_text_size = 12,
  sig_bar_nudge = 0.06,
  sig_star_nudge = sig_bar_nudge * 0.75,
  sig_text_nudge = sig_bar_nudge * 1.35,
  step_increase = if (grepl("^log", y_scale)) 0 else sig_bar_nudge + 0.05,
  sig_method = "p_by_normality",
  p_case = "upper",
  p_spaces = TRUE,
  n_breaks = 4,
  breaks_fn = pretty,
  y_max = NULL,
  y_min = 0,
  expand_y = 0,
  expand_x = waiver(),
  censor_fn = rescale_none,
  show_all = TRUE,
  seed = 1234,
  ...
)
}
\arguments{
\item{df}{Data frame in long format containing continuous variable, categorical or integer variable, and grouping variable}

\item{formula}{y ~ x format}

\item{grouping_var}{Variable to group by. Enter as quoted or unquoted variable name}

\item{grouping_var_order}{Order for levels of \code{grouping_var} along x axis. Enter as character vector}

\item{rev_grouping_var_order}{If \code{TRUE}, levels of \code{grouping_var} are reversed}

\item{x_order}{Order of x values. Enter as character vector. Default is \code{NULL}}

\item{rev_x_order}{If \code{FALSE} (default), order of x axis variable is not reversed}

\item{x, y}{Variables for x  and y axis, respectively. Enter as quoted or unquoted variable names}

\item{beeswarm}{If \code{FALSE} (default), jittered points. If \code{TRUE}, beeswarm plot}

\item{beeswarm_method}{Method for \code{geom_quasirandom} function. Default is \code{"smiley"}}

\item{width, bar_width}{Width of bars. Default is 0.2*number of x values}

\item{beeswarm_width}{Width argument for beeswarm points. Default is \code{0.15}}

\item{n_bins}{Number of bins. Used in \code{ggbeeswarm::geom_quasirandom} function. Default uses 30/number of points}

\item{band_width}{Bandwidth. Used in \code{ggbeeswarm::geom_quasirandom} function. Default is \code{1 + n_bins/5}}

\item{dodge}{Entry for position_dodge. Default is \code{0.7}}

\item{jitter_width}{Width between groups of points. Default is \code{0.2}}

\item{colors}{Alias for \code{point_colors}}

\item{color_var}{Variable used to color points. Default uses \code{grouping_var.} Enter as quoted or unquoted variable name}

\item{point_shapes}{Point shapes. Options: \code{"circle"} (default), \code{"square"}, \code{"triangle"}, \code{"diamond"}. Enter as quoted shape}

\item{point_size}{Point size in pts units. Default is \code{5}}

\item{show_points}{If \code{TRUE} (default), points are displayed. If \code{FALSE}, points not displayed in plot}

\item{small_points}{If \code{TRUE}, small points are plotted in 1 color for all groups. If \code{FALSE}, plots are pointed as usual}

\item{point_alpha}{Transparency of point colors. Enter as numeric 0-1. Default is \code{0.9}}

\item{bar_alpha}{Alpha value for bar fill. Enter as length 1 numeric. Default is \code{0}}

\item{bar_border_thickness}{Thickness of line surrounding bars in pts. Enter as length 1 numeric. Default is \code{0.75}}

\item{point_border_thickness}{Thickness of line surrounding points in pts units. Default is \code{1}}

\item{bar_border_color}{Color for bar border. Default is \code{"black"}}

\item{point_shape_var}{Variable used to determine point shapes. Default uses \code{grouping_var.} Enter as quoted or unquoted variable name}

\item{show_errorbar}{If \code{TRUE}, error bars are displayed}

\item{summary_fn}{Function used to determine midpoint of error bars. Default is \code{Mean}}

\item{error_fn}{Function used to determine ends of error bars. Default is \code{SE}}

\item{errorbar_width_multiplier}{Multiplied by \code{bar_width} to determine width of error bars. Enter as length 1 numeric. Default is \code{0.5}}

\item{show_legend}{If \code{FALSE} (default), legend is not displayed}

\item{show_legend_points}{If \code{FALSE} (default) legend for points is not displayed}

\item{show_legend_bars}{If \code{FALSE} (default) legend for bars is not displayed}

\item{legend_title}{Legend title. Default is blank}

\item{y_scale}{Scaling for y axis. Options: \code{"regular"} (default), \code{"log"}, \code{"scientific"}}

\item{y_axis_breaks}{Numeric vector or function specifying location of ticks along y axis}

\item{y_axis_labels}{Vector or function specifying y axis tick labels}

\item{x_axis_breaks}{Numeric vector or function specifying location of ticks along x axis}

\item{x_axis_labels}{Vector or function specifying x axis tick labels}

\item{y_title}{Alias for \code{y_axis_title}}

\item{y_axis_title}{Enter as quoted string}

\item{x_title}{Alias for \code{x_axis_title}}

\item{x_axis_title}{Enter as quoted string}

\item{plot_title}{Text to be used as main plot title. Enter as quoted string. Default is \code{""} (i.e. no title)}

\item{show_sig}{If \code{TRUE} (default), significance annotation displayed. If \code{FALSE}, no significance annotation displayed}

\item{stars}{If \code{TRUE} (default), significance stars are displayed. If \code{FALSE}, actual P values are displayed}

\item{sig_font_color}{Color for text in significance annotations. Default is \code{"black"}}

\item{sig_bar_color}{Color for significance bars and annotations. Default is \code{"black"}}

\item{sig_bar_thickness}{Line thickness for significance bars in linewidth units. Default is \code{0.6}}

\item{show_ns}{If \code{TRUE} (default), non-significant annotations are displayed. If \code{FALSE}, no annotation is displayed comparisons in which P >= 0.05}

\item{ns_symbol}{Symbol to use when P >= 0.05. Only relevant when \code{show_ns = TRUE}. Default is \code{"ns"}}

\item{sig_star_size}{Size of significance stars in pts. Only relevant when \code{stars = TRUE}. Default is \code{20}}

\item{sig_text_size}{Font size for P values in pts. Only relevant when \code{stars = FALSE}. Default is \code{12}}

\item{sig_bar_nudge}{Nudge factor along y axis for significance bars. Distance between point or bar and significance bar is sig_bar_nudge x maximum y value on plot. Default is \code{0.08}}

\item{sig_star_nudge}{Nudge factor along y axis for significance stars. Distance between point or bar and significance stars is \code{sig_star_nudge}*maximum y value on plot. Default is \code{0.05}}

\item{sig_text_nudge}{Nudge factor along y axis for significance annotation text. Distance between point or bar and significance annotation text is \code{sig_text_nudge}*maximum y value on plot. Default is \code{0.09}}

\item{step_increase}{Step increase as proportion of view port. Default is \code{0.11}}

\item{sig_method}{Options: \code{"p_by_normality"} (default), \code{"mann_whitney"}, \code{"t_test_welch"}, \code{"t_test_no_welch"}. Enter as quoted test type}

\item{p_case}{Case of "p" used in significance annotations when \code{stats = FALSE.} Options: \code{"upper"} ("P", default) or \code{"lower"} ("p"). Enter as quoted text}

\item{p_spaces}{If \code{TRUE} (default), spaces placed between p, equality symbol, and P-value in significance annotation. If \code{FALSE}, no space between "p", equality symbol, and P-value in significance annotation. Only relevant when \code{stars = FALSE}}

\item{n_breaks}{Number of tick marks for y axis. Default is \code{3}}

\item{breaks_fn}{Function used to generate y axis breaks. Passed to \code{.create_axis_breaks}. Default is \code{pretty}}

\item{y_max}{Maximum y axis value. Default is \code{NULL}}

\item{y_min}{Minimum value to use for y axis. Default is \code{0}}

\item{expand_y}{Expansion of y axis around y = 0. Default is \code{0}}

\item{expand_x}{Argument entered for expand argument of \code{scale_x_continuous}. Default is \code{waiver()}}

\item{censor_fn}{Function used to transform data outside y axis limits. Default is \code{rescale_none.} Alternative: \code{scales::censor}}

\item{show_all}{If \code{TRUE} (default), all significance bars are displayed}

\item{seed}{Length 1 integer vector to make point jittering reproducible. Only relevant when \code{beeswarm = FALSE}. Default is \code{1234}}

\item{...}{Arguments passed to \code{theme_fn()}}
}
\value{
ggplot object
}
\description{
Bar plot with overlying points
}
