% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_box.R
\name{plot_box}
\alias{plot_box}
\title{Box plot}
\usage{
plot_box(
  df,
  formula,
  grouping_var = NULL,
  x = NULL,
  y = NULL,
  grouping_var_order = NULL,
  rev_grouping_var_order = FALSE,
  x_order = NULL,
  color_var = NULL,
  colors = c("#0072B5", "#BC3C29", "#999999", "#333333", "#CC79A7", "#56B4E9", "#E69F00",
    "#20854E", "#F0E442", "#7D5FA7", "#97572B", "#005E7A", "#00A0B0", "#8D9D58",
    "#DA3978", "#61D04F", "#FBA27D", "#FFD500", "#00C5CD", "#F7F7F7"),
  alpha = 1,
  border_thickness = 0.75,
  border_colors = "black",
  width = NULL,
  errorbar_thickness = border_thickness,
  errorbar_width_multiplier = 0.5,
  show_points = FALSE,
  point_shapes = "circle",
  point_shape_var = NULL,
  point_color_var = NULL,
  point_colors = "#2A2D34",
  point_border_color = "black",
  point_border_thickness = 0,
  point_size = 0.5,
  jitter_width = 0.2,
  show_legend = FALSE,
  legend_title = "",
  y_scale = "regular",
  y_axis_breaks = NULL,
  y_axis_labels = NULL,
  y_axis_title = "",
  x_title = NULL,
  x_axis_title = x_title,
  plot_title = "",
  dodge = 0.7,
  n_breaks = 3,
  y_max = NULL,
  y_min = NULL,
  expand_y = 0.1,
  x_axis_labels = waiver(),
  censor_fn = rescale_none,
  base_size = 14,
  theme_fn = theme_custom,
  ...
)
}
\arguments{
\item{df}{Data frame in long format containing continuous variable, categorical or integer variable, and grouping variable}

\item{formula}{y ~ x format}

\item{grouping_var}{Variable to group by. Enter as quoted or unquoted variable name}

\item{x, y}{Variables for x  and y axis, respectively. Enter as quoted or unquoted variable names}

\item{grouping_var_order}{Order for levels of \code{grouping_var} along x axis. Enter as character vector}

\item{rev_grouping_var_order}{If \code{TRUE}, levels of \code{grouping_var} are reversed}

\item{x_order}{Order of x values. Enter as character vector. Default is \code{NULL}}

\item{color_var}{Variable used to color boxes. Enter as quoted or unquoted variable name}

\item{colors}{Hexadecimal or quoted color names. Enter using \code{c()}. Default uses black/red}

\item{alpha}{Alpha value for box fill. Default is \code{1}}

\item{border_thickness}{Thickness of line surrounding boxes in pt units. Default is \code{0.75}}

\item{border_colors}{Color of line surrounding boxes. Default is \code{"black"}}

\item{width}{Width of boxes. Default is \code{0.5}}

\item{errorbar_thickness}{Thickness of errorbar lines. Default is \code{border_thickness}}

\item{errorbar_width_multiplier}{Multiplied by \code{bar_width} to determine width of error bars. Enter as length 1 numeric. Default is \code{0.5}}

\item{show_points}{If \code{TRUE}, points are displayed. Default is \code{FALSE}}

\item{point_shapes}{Point shapes. Options: \code{"circle"} (default), \code{"square"}, \code{"triangle"}, \code{"diamond"}. Enter as quoted shape}

\item{point_shape_var}{Variable used to determine point shapes. Default uses \code{grouping_var.} Enter as quoted or unquoted variable name}

\item{point_color_var}{Variable used to determine point colors. Enter as quoted or unquoted variable name. Default uses \code{grouping_var}}

\item{point_colors}{Point fill colors. Enter as character vector of color names or hexadecimal codes}

\item{point_border_thickness}{Thickness of line surrounding points in pts units. Default is \code{1}}

\item{point_size}{Point size in pts units. Default is \code{5}}

\item{jitter_width}{Width between groups of points. Default is \code{0.2}}

\item{show_legend}{If \code{FALSE} (default), legend is not displayed}

\item{legend_title}{Legend title. Default is blank}

\item{y_scale}{Scaling for y axis. Options: \code{"regular"} (default), \code{"log"}, \code{"scientific"}}

\item{y_axis_breaks}{Numeric vector or function specifying location of ticks along y axis}

\item{y_axis_labels}{Vector or function specifying y axis tick labels}

\item{y_axis_title}{Enter as quoted string}

\item{x_title}{Alias for \code{x_axis_title}}

\item{x_axis_title}{Enter as quoted string}

\item{plot_title}{Text to be used as main plot title. Enter as quoted string. Default is \code{""} (i.e. no title)}

\item{dodge}{Entry for position_dodge. Default is \code{0.7}}

\item{n_breaks}{Number of tick marks for y axis. Default is \code{3}}

\item{y_max}{Maximum y axis value. Default is \code{NULL}}

\item{y_min}{Minimum value to use for y axis. Default is \code{0}}

\item{expand_y}{Expansion of y axis around y = 0. Default is \code{0}}

\item{x_axis_labels}{Vector or function specifying x axis tick labels}

\item{censor_fn}{Function used to transform data outside y axis limits. Default is \code{rescale_none.} Alternative: \code{scales::censor}}

\item{base_size}{Font size used in theme}

\item{theme_fn}{Theme function. Default is \code{theme_custom}}

\item{...}{Arguments passed to \code{theme_fn()}}
}
\description{
Box plot
}
