% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_correlation.R
\name{plot_cor}
\alias{plot_cor}
\title{Plot correlation coefficient, p value, linear regression line}
\usage{
plot_cor(
  df,
  formula = NULL,
  x = NULL,
  y = NULL,
  grouping_var = NULL,
  grouping_var_order = NULL,
  rev_grouping_var_order = FALSE,
  point_color_var = NULL,
  colors = c("#0072B5", "#BC3C29", "#868686", "#2A2D34", "#CC79A7", "#56B4E9", "#E69F00",
    "#20854E", "#F0E442", "#7D5FA7", "#97572B", "#005E7A", "#00A0B0", "#8D9D58",
    "#DA3978", "#61D04F", "#FBA27D", "#FFD500", "#00C5CD", "#F7F7F7"),
  point_colors = colors,
  point_alpha = 0.9,
  point_shape_var = NULL,
  point_shapes = "circle",
  point_size = 4,
  point_border_thickness = 0.5,
  point_border_colors = "#333333",
  line_color_var = NULL,
  line_colors = point_colors,
  line_thickness = 1,
  show_error = TRUE,
  error_colors = "#999999",
  show_legend = FALSE,
  y_scale = "regular",
  y_axis_breaks = NULL,
  y_max = NULL,
  y_min = NULL,
  expand_y = 0.1,
  y_axis_labels = waiver(),
  y_axis_title = waiver(),
  x_scale = "regular",
  x_axis_breaks = NULL,
  x_max = NULL,
  x_min = NULL,
  expand_x = 0.1,
  x_axis_labels = waiver(),
  x_axis_title = waiver(),
  plot_title = NULL,
  show_corr_coef = TRUE,
  anno_position = "best",
  method = "spearman",
  corr_prefix = NULL,
  p_prefix = "P",
  show_n = TRUE,
  n_prefix = "n",
  show_rsq = TRUE,
  point_geom_fn = ggplot2::geom_point,
  ...
)
}
\arguments{
\item{df}{Data frame in long format containing continuous variable, categorical or integer variable, and grouping variable}

\item{formula}{y ~ x format}

\item{x, y}{Continuous variables variables. Enter as quoted or unquoted variable names}

\item{grouping_var}{Variable to group by. Enter as quoted or unquoted variable name}

\item{rev_grouping_var_order}{If \code{FALSE} (default), levels of grouping_var ordered alphabetically. If \code{TRUE}, order of grouping_var levels is reversed}

\item{point_color_var}{Variable used to determine point colors. Enter as quoted or unquoted variable name. Default uses \code{grouping_var}}

\item{colors}{Alias for \code{point_colors}}

\item{point_colors}{Point fill colors. Enter as character vector of color names or hexadecimal codes}

\item{point_alpha}{Transparency of point colors. Enter as numeric 0-1. Default is \code{0.9}}

\item{point_shape_var}{Variable used to determine point shapes. Default uses \code{grouping_var.} Enter as quoted or unquoted variable name}

\item{point_size}{Point size in pts units. Default is \code{5}}

\item{point_border_thickness}{Thickness of line surrounding points in pts units. Default is \code{1}}

\item{line_colors}{Color for lines. Hexadecimal or quoted color names. Default is \code{"black"}}

\item{line_thickness}{Thickness of line in mm. Default is \code{0.75}}

\item{show_error}{If \code{TRUE} (default), error region displayed with color as determined by \code{error_colors}. If \code{FALSE}, error region not displayed}

\item{error_colors}{Color for error region around trend line. Only relevant when \code{show_error = TRUE}}

\item{show_legend}{If \code{FALSE} (default), legend is not displayed}

\item{y_scale}{Scaling for y axis. Options: \code{"regular"} (default), \code{"log"}, \code{"scientific"}}

\item{y_axis_breaks}{Numeric vector or function specifying location of ticks along y axis}

\item{y_max}{Maximum y axis value. Default is \code{NULL}}

\item{y_min}{Minimum value to use for y axis. Default is \code{0}}

\item{expand_y, expand_x}{Expansion to add to y and x axis, respectively. Default is \code{0.1}}

\item{y_axis_labels}{Vector or function specifying y axis tick labels}

\item{y_axis_title}{Enter as quoted string}

\item{x_scale}{Scale used for x axis. Options: \code{"regular"} (default), \code{"log"}, \code{"scientific"}}

\item{x_axis_breaks}{Numeric vector or function specifying location of ticks along x axis}

\item{x_max}{Maximum value for x axis. Default is \code{NULL}}

\item{x_min}{Minimum value for x axis. Default is \code{NULL}}

\item{x_axis_labels}{Vector or function specifying x axis tick labels}

\item{x_axis_title}{Enter as quoted string}

\item{plot_title}{Text to be used as main plot title. Enter as quoted string. Default is \code{""} (i.e. no title)}

\item{show_corr_coef}{If \code{TRUE} (default), correlation coefficient and P value are displayed on plot}

\item{anno_position}{Position of correlation coefficient and P value annotation. Can enter as numeric vector of length 2 containing x and y coordinates, \code{"best"}, a length 1 character vector with some combination of the following letters: \code{"t"}, \code{"b"}, \code{"l"}, \code{"r"}}

\item{method}{Method for performing correlation test. Options: \code{"spearman"} (default), \code{"pearson"}, \code{"kendall"}}

\item{corr_prefix}{Prefix for correlation coefficient annotation. If \code{NULL}, rho used for Spearman, R for Pearson, tau for Kendall}

\item{p_prefix}{Prefix for P values. Enter as length 1 character vector. Default is \code{"P"}}

\item{show_n}{If \code{TRUE} (default), number of observations displayed as component of correlation annotation}

\item{n_prefix}{Prefix for number of observations. Enter as length 1 character vector. Default is \code{"n"}}

\item{show_rsq}{If \code{TRUE} (default), R squared and P value are displayed}

\item{point_geom_fn}{Geom function to generate points. Default is \code{ggplot2::geom_point}. Alternative: \code{scattermore::geom_scattermore}}

\item{...}{Arguments passed to \code{theme_fn()}}
}
\value{
ggplot object
}
\description{
Plot correlation coefficient, p value, linear regression line
}
