% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_other.R
\name{plot_ex}
\alias{plot_ex}
\title{Example plot - y vs. x}
\usage{
plot_ex(fn = "plot_bar_point", ...)
}
\arguments{
\item{fn}{Plot function. Default is \code{plot_bar_point}}

\item{...}{Arguments passed to \code{fn}}

\item{formula}{formula. Default is \code{re ~ day}}

\item{group}{Grouping variable. Default is \code{"genotype"}}
}
\value{
ggplot object
}
\description{
Example plot - y vs. x
}
