% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cowplot.R
\name{plot_grid}
\alias{plot_grid}
\alias{align_plots}
\title{Arrange multiple plots into a grid}
\usage{
plot_grid(
  ...,
  plotlist = NULL,
  align = c("none", "h", "v", "hv"),
  axis = c("none", "l", "r", "t", "b", "lr", "tb", "tblr"),
  nrow = NULL,
  ncol = NULL,
  byrow = TRUE,
  scale = 1,
  rel_widths = 1,
  rel_heights = 1,
  labels = NULL,
  label_size = 14,
  label_fontfamily = NULL,
  label_fontface = "bold",
  label_colour = NULL,
  label_x = 0,
  label_y = 1,
  hjust = -0.5,
  vjust = 1.5,
  greedy = TRUE
)

align_plots(
  ...,
  plotlist = NULL,
  align = c("none", "h", "v", "hv"),
  axis = c("none", "l", "r", "t", "b", "lr", "tb", "tblr"),
  greedy = TRUE
)
}
\arguments{
\item{...}{Comma separated list of plots}

\item{plotlist}{List of plots}

\item{align}{Options: \code{"none"} (default), \code{"horizontal"}, \code{"vertical"}, \code{"hv"} (both)}

\item{axis}{Options: \code{"none"} (default), \code{"left"}, \code{"top"}, \code{"bottom"}, \code{"right"}, or any combination (i.e. "tbrl")}

\item{nrow, ncol}{Number of rows and columns, respectively, in plot grid}

\item{byrow}{If \code{TRUE} (default), plots arranged by row. If \code{FALSE}, plots arranged by column}

\item{scale}{Scaling factor for plot. Default is \code{1}}

\item{rel_heights, rel_widths}{Relative column heights and widths, respectively, in plot grid. \code{rel_widths = c(2, 1)} will make the 1st column 2x width of 2nd column}

\item{labels}{List of labels to be added to plots. Enter \code{labels = "AUTO"} for upper case and \code{labels = "auto"} for lower case}

\item{label_size}{Font size for labels. Default is \code{14}}

\item{label_fontfamily}{Font family for labels}

\item{label_fontface}{Font face for labels}

\item{label_colour}{Font color for labels}

\item{label_x, label_y}{x and y positions respectively for labels relative to the given subplot. Numeric 0-1}

\item{hjust}{Adjust horizontal position of each label. More negative values move the label further to the right on the plot canvas. Can be a single value (applied to all labels) or a vector of values (one for each label). Default is \code{-0.5}}

\item{vjust}{Adjust vertical position of each label. More positive values move the label further down on the plot canvas. Can be a single value (applied to all labels) or a vector of values (one for each label). Default is \code{1.5}}

\item{greedy}{Adjustment method for plot alignment}
}
\value{
ggplot
}
\description{
Used by \code{plot_grid}
}
