% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Heatmap of summarized numeric}
\usage{
plot_heatmap(
  df,
  row_var = NULL,
  col_var = NULL,
  scale_by = c("rows", "columns", "none"),
  cluster_cols = FALSE,
  cluster_rows = TRUE,
  size = 15,
  cellheight = size,
  cellwidth = size,
  colors = c("#0072B5", "white", "#BC3C29"),
  n_colors = 100,
  border_colors = "black",
  show_colnames = TRUE,
  show_rownames = TRUE,
  font_size = 18,
  font_size_rownames = font_size,
  font_size_colnames = font_size,
  rowname_location = "right",
  rowname_align = "right",
  colname_location = "bottom",
  colname_angle = 90,
  colname_align = "right",
  col_gaps = 0,
  row_gaps = 0,
  treeheight_col = 0,
  treeheight_row = 0,
  show_legend = TRUE,
  mat = df,
  legend = show_legend,
  fontsize = font_size,
  fontsize_row = font_size_rownames,
  fontsize_col = font_size_colnames,
  angle_col = colname_angle,
  gaps_col = col_gaps,
  gaps_row = row_gaps
)
}
\arguments{
\item{df, mat}{Data frame with column containing rownames}

\item{row_var}{Variable used for rows. Enter as quoted or unquoted row name}

\item{scale_by}{Scaling of rows and columns. Options: \code{"rows" }(default), \code{"columns"}, \code{"none"}}

\item{cluster_cols}{If \code{TRUE}, columns are clustered. Default is \code{FALSE}}

\item{cluster_rows}{If \code{TRUE} (default), rows are clustered}

\item{size, cellheight, cellwidth}{Height and width of heatmap squares in pts. Default is \code{15}}

\item{colors}{Colors used to generate heatmap colors. Enter 2-3 colors. Default is blue and red}

\item{n_colors}{Number of unique colors. Default is \code{100}}

\item{border_colors}{Color of line surrounding cells in heatmap. Default is \code{"black"}}

\item{show_colnames}{If \code{TRUE} (default), column names are displayed on heatmap}

\item{show_rownames}{If \code{TRUE} (default), row names are displayed on heatmap}

\item{font_size, fontsize, font_size_rownames, fontsize_row, font_size_colnames, fontsize_col}{Font size for row and column labels in heatmap. Default is \code{18}}

\item{rowname_location}{Location of row labels. Options include \code{"left"} (default), \code{"right"}, \code{"none"}}

\item{rowname_align}{Horizontal alignment of row labels. Options include \code{"right"} (default), \code{"left"}, \code{"center"}}

\item{colname_location}{Location of column labels. Options include \code{"top"} (default), \code{"bottom"}, \code{"none"}}

\item{colname_angle, angle_col}{Angle for column labels. Enter as length 1 numeric. Default is \code{90}}

\item{colname_align}{Vertical alignment of column labels. Options include \code{"bottom"} (default), \code{"top"}, \code{"center"}}

\item{row_gaps, gaps_row, col_gaps, gaps_col}{Row numbers to separate heatmap rows and columns, respectively. Enter as length 1 numeric. Default uses no gaps for either}

\item{treeheight_row, treeheight_col}{Size of dendrogram for rows and columns, respectively. Default is \code{0} for both}

\item{show_legend, legend}{If \code{TRUE} (default), legend is displayed}
}
\value{
pheatmap object with ggplot object added
}
\description{
Heatmap of summarized numeric
}
