% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_line.R
\name{plot_line}
\alias{plot_line}
\title{Plot lines}
\usage{
plot_line(
  df,
  formula = NULL,
  grouping_var = NULL,
  grouping_var_order = NULL,
  rev_grouping_var_order = FALSE,
  x = NULL,
  y = NULL,
  colors = c("#0072B5", "#BC3C29", "#868686", "#2A2D34", "#CC79A7", "#56B4E9", "#E69F00",
    "#20854E", "#F0E442", "#7D5FA7", "#97572B", "#005E7A", "#00A0B0", "#8D9D58",
    "#DA3978", "#61D04F", "#FBA27D", "#FFD500", "#00C5CD", "#F7F7F7"),
  line_type = "solid",
  line_type_var = NULL,
  line_thickness = 1,
  line_color_var = NULL,
  line_colors = colors,
  show_points = TRUE,
  point_color_var = line_color_var,
  point_shape_var = NULL,
  point_shapes = "circle",
  point_size = 3,
  point_colors = colors,
  point_border_color = "black",
  point_border_thickness = 0,
  show_legend = FALSE,
  y_scale = "regular",
  x_axis_title = waiver(),
  y_axis_breaks = NULL,
  y_axis_labels = NULL,
  y_axis_title = waiver(),
  x_axis_breaks = waiver(),
  x_axis_labels = waiver(),
  expand_y = 0.05,
  expand_x = waiver(),
  plot_title = NULL,
  n_breaks = 3,
  y_min = NULL,
  y_max = NULL,
  censor_fn = rescale_none,
  legend_title = NULL,
  theme_fn = theme_custom,
  ...
)
}
\arguments{
\item{df}{Data frame in long format containing continuous variable (y), categorical or integer variable (x), and grouping variable (grouping_var)}

\item{formula}{y ~ x format}

\item{grouping_var}{Variable to group by. Enter as quoted or unquoted variable name}

\item{rev_grouping_var_order}{If \code{FALSE} (default), levels of grouping_var ordered alphabetically. If \code{TRUE}, order of grouping_var levels is reversed}

\item{x, y}{Variables for x and y axis respectively. Enter as quoted or unquoted variable names}

\item{colors}{Color used for points and lines}

\item{line_type}{Type of line. Options include \code{"solid"} (default), \code{"dashed"}, \code{"longdash"}, \code{"twodash"}, \code{"dotdash"}, \code{"dotted"}. Enter as quoted linetype}

\item{line_thickness}{Thickness of line in mm. Default is \code{0.75}}

\item{line_colors}{Color for lines. Hexadecimal or quoted color names. Default is \code{"black"}}

\item{show_points}{If \code{TRUE} (default), points are displayed}

\item{point_size}{Point size. Default is \code{3}}

\item{point_colors}{Fill color for points. Hexadecimal or quoted color names. Enter using \code{c()}}

\item{point_border_thickness}{Thickness of line surrounding points in pt units. Default is \code{1}}

\item{show_legend}{If \code{TRUE}, legend displayed. If \code{FALSE}, legend not displayed}

\item{y_scale}{Scaling for y-axis. Options: \code{"regular"} (default), \code{"log"}, \code{"scientific"}}

\item{x_axis_title}{Enter as quoted string}

\item{y_axis_breaks}{Numeric vector or function specifying location of ticks along y axis}

\item{y_axis_labels}{Vector or function specifying y axis tick labels}

\item{y_axis_title}{Enter as quoted string}

\item{x_axis_breaks}{Numeric vector or function specifying location of ticks along x axis}

\item{x_axis_labels}{Vector or function specifying x axis tick labels}

\item{expand_y}{Expansion of y axis around y = 0. Default is \code{0}}

\item{expand_x}{Argument entered for \code{expand} argument of \code{scale_x_continuous}}

\item{plot_title}{Text to be used as main plot title. Enter as quoted string. Default is \code{""} (i.e. no title)}

\item{n_breaks}{Number of tick marks for y axis. Default is \code{3}}

\item{y_min}{Minimum value to use for y axis Default is \code{0}}

\item{y_max}{Maximum y axis value. Default is \code{max(df$y)}}

\item{censor_fn}{Function used to transform data outside y axis limits. Default is \code{rescale_none.} Alternative: \code{scales::censor}}

\item{...}{Arguments passed to \code{theme_custom()}}
}
\description{
Plot lines
}
