% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_other.R
\name{plot_map}
\alias{plot_map}
\title{Plot of world map}
\usage{
plot_map(
  countries = NULL,
  color_for_countries = "#C8543B",
  color_other_countries = "#F7F7F7",
  line_color = "#333333",
  line_thickness = 0.3,
  ...
)
}
\arguments{
\item{countries}{Countries to highlight on map. Enter as character vector}

\item{color_for_countries}{Color used to fill countries mentioned in countries argument}

\item{color_other_countries}{Color used to fill countries not mentioned in countries argument. Default is \code{"grey"}}

\item{line_color}{Line color for country borders. Default is \code{"black"}}

\item{line_thickness}{Line thickness for country borders. Default is \code{0.3}}

\item{...}{Arguments passed to \code{theme_blank}}
}
\value{
ggplot object
}
\description{
Plot of world map
}
