% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_perc.R
\name{plot_perc}
\alias{plot_perc}
\title{Plot percentage +/- CI for each group}
\usage{
plot_perc(
  df,
  formula = NULL,
  grouping_var = NULL,
  x = NULL,
  y = NULL,
  dodge = 0.7,
  x_order = NULL,
  rev_x_order = FALSE,
  grouping_var_order = NULL,
  rev_grouping_var_order = FALSE,
  point_shapes = "square",
  point_shape_var = NULL,
  point_color_var = NULL,
  colors = c("#333333", "#0072B5", "#BC3C29", "#868686", "#CC79A7", "#56B4E9", "#E69F00",
    "#20854E", "#F0E442", "#7D5FA7", "#97572B", "#005E7A", "#00A0B0", "#8D9D58",
    "#DA3978", "#61D04F", "#FBA27D", "#FFD500", "#00C5CD", "#F7F7F7"),
  point_colors = colors,
  point_size = 3,
  point_border_thickness = 1,
  point_border_colors = "black",
  show_errorbar = TRUE,
  errorbar_width = 0.25,
  errorbar_thickness = 0.75,
  show_legend = FALSE,
  legend_title = "",
  as_perc = TRUE,
  y_title = waiver(),
  y_axis_title = y_title,
  invert = FALSE,
  y_axis_breaks = NULL,
  y_axis_labels = NULL,
  breaks_fn = pretty,
  n_breaks = 4,
  y_min = NULL,
  y_max = NULL,
  expand_y = 0.1,
  censor_fn = rescale_none,
  x_title = NULL,
  x_axis_label_angle = 45,
  x_angle = x_axis_label_angle,
  x_axis_title = x_title,
  x_axis_labels = waiver(),
  expand_x = waiver(),
  geom = "point",
  plot_title = NULL,
  theme_fn = theme_custom,
  ...
)
}
\arguments{
\item{df}{Data frame in long format containing continuous variable, categorical or integer variable, and grouping variable}

\item{formula}{y ~ x format}

\item{grouping_var}{Variable to group by. Enter as quoted or unquoted variable name}

\item{x, y}{Variables for x  and y axis, respectively. Enter as quoted or unquoted variable names}

\item{dodge}{Entry for position_dodge. Default is \code{0.7}}

\item{x_order}{Order of x values. Enter as character vector. Default is \code{NULL}}

\item{rev_x_order}{If \code{FALSE} (default), order of x axis variable is not reversed}

\item{grouping_var_order}{Order for levels of \code{grouping_var} along x axis. Enter as character vector}

\item{rev_grouping_var_order}{If \code{TRUE}, levels of \code{grouping_var} are reversed}

\item{point_shapes}{Point shapes. Options: \code{"circle"} (default), \code{"square"}, \code{"triangle"}, \code{"diamond"}. Enter as quoted shape}

\item{point_shape_var}{Variable used to determine point shapes. Default uses \code{grouping_var.} Enter as quoted or unquoted variable name}

\item{point_color_var}{Variable used to determine point colors. Enter as quoted or unquoted variable name. Default uses \code{grouping_var}}

\item{colors}{Alias for \code{point_colors}}

\item{point_colors}{Point fill colors. Enter as character vector of color names or hexadecimal codes}

\item{point_size}{Point size in pts units. Default is \code{5}}

\item{point_border_thickness}{Thickness of line surrounding points in pts units. Default is \code{1}}

\item{point_border_colors}{Color of line surrounding points. Default is \code{"black"}}

\item{show_errorbar}{If \code{TRUE}, error bars are displayed}

\item{errorbar_width}{Width of error bars. Default is \code{n_unique(df$x)/10}}

\item{errorbar_thickness}{Line thickness of error bar. Default is \code{0.75}}

\item{show_legend}{If \code{FALSE} (default), legend is not displayed}

\item{legend_title}{Legend title. Default is blank}

\item{as_perc}{If \code{TRUE} (default), y axis is plotted as percentage (0-100). If \code{FALSE}, y axis plotted as proportion (0-1)}

\item{y_title}{Alias for \code{y_axis_title}}

\item{y_axis_title}{Enter as quoted string}

\item{invert}{If \code{TRUE}, 1 - proportion (or 100 - percentage) plotted along y axis}

\item{y_axis_breaks}{Numeric vector or function specifying location of ticks along y axis}

\item{y_axis_labels}{Vector or function specifying y axis tick labels}

\item{breaks_fn}{Function used to generate y axis breaks. Passed to \code{.create_axis_breaks}. Default is \code{pretty}}

\item{n_breaks}{Number of tick marks for y axis. Default is \code{3}}

\item{y_min}{Minimum value to use for y axis. Default is \code{0}}

\item{y_max}{Maximum y axis value. Default is \code{NULL}}

\item{expand_y}{Expansion of y axis around y = 0. Default is \code{0}}

\item{censor_fn}{Function used to transform data outside y axis limits. Default is \code{rescale_none.} Alternative: \code{scales::censor}}

\item{x_title}{Alias for \code{x_axis_title}}

\item{x_axis_title}{Enter as quoted string}

\item{x_axis_labels}{Vector or function specifying x axis tick labels}

\item{expand_x}{Argument entered for expand argument of \code{scale_x_continuous}. Default is \code{waiver()}}

\item{geom}{geom function to plot percent. Default is \code{"point"}}

\item{plot_title}{Text to be used as main plot title. Enter as quoted string. Default is \code{""} (i.e. no title)}

\item{theme_fn}{Theme function. Default is \code{theme_custom}}

\item{...}{Arguments passed to \code{theme_fn()}}
}
\value{
ggplot object
}
\description{
Plot percentage +/- CI for each group
}
