% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_other.R
\name{plot_piechart}
\alias{plot_piechart}
\title{Pie chart}
\usage{
plot_piechart(
  df,
  categorical_var,
  count_var = "n",
  rev_order = FALSE,
  order_by = count_var,
  fill_colors = c("#333333", "#00A1D5", "#E41A1C", "#6761A8", "#009872", "#EFC000",
    "#045A8D", "#FF7F00", "#BABABA", "#FFFFFF"),
  colors = fill_colors,
  border_color = "#000000",
  show_legend = TRUE,
  legend_title = ""
)
}
\arguments{
\item{df}{Data frame}

\item{categorical_var}{Variable in \code{df} containing categories. Enter as quoted column name}

\item{count_var}{Variable in \code{df} containing counts (or proportions or percentages) for each \code{categorical_var}. Enter as quoted column name. Default is \code{"n"}}

\item{rev_order}{If \code{TRUE}, order of levels in pie chart are reversed so that levels are plotted from most to least frequent. Default is \code{FALSE} (levels are plotted from least frequent to most frequent)}

\item{order_by}{Variable in \code{df} to order unique values in \code{df[[categorical_var]]}. Enter as quoted column column name. Default is \code{count_var}. If \code{NULL}, values appear in piechart in same order as row order in \code{df}. In this case, \code{rev_order} is ignored.}

\item{fill_colors}{Pie colors for each unique value in \code{df[[categorical_var]]}}

\item{border_color}{Line color for pie borders. Default is black}

\item{show_legend}{If \code{TRUE} (default), legend is displayed}

\item{legend_title}{Legend title. Default is blank}
}
\value{
ggplot object
}
\description{
Pie chart
}
