% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sina.R
\name{plot_sina}
\alias{plot_sina}
\title{Sina plot}
\usage{
plot_sina(
  df,
  formula,
  grouping_var = NULL,
  grouping_var_order = NULL,
  rev_grouping_var_order = FALSE,
  x_order = NULL,
  fn_x_order = mean,
  rev_x_order = FALSE,
  x = NULL,
  y = NULL,
  colors = c("#0072B5", "#BC3C29", "#999999", "#333333", "#CC79A7", "#56B4E9", "#E69F00",
    "#20854E", "#F0E442", "#7D5FA7", "#97572B", "#005E7A", "#00A0B0", "#8D9D58",
    "#DA3978", "#61D04F", "#FBA27D", "#FFD500", "#00C5CD", "#F7F7F7"),
  point_color_var = NULL,
  point_shape_var = NULL,
  point_shapes = c("circle", "circle"),
  point_border_color = "black",
  alpha = 0.9,
  point_alpha = alpha,
  width = 0.2,
  dodge = 0.7,
  point_border_thickness = 0.75,
  show_legend = FALSE,
  legend_title = "",
  y_scale = "regular",
  y_title = waiver(),
  y_axis_title = y_title,
  x_title = waiver(),
  x_axis_title = x_title,
  plot_title = NULL,
  n_breaks = 3,
  breaks_fn = pretty,
  x_breaks = NULL,
  x_axis_breaks = x_breaks,
  x_labels = NULL,
  x_axis_labels = x_labels,
  y_breaks = NULL,
  y_axis_breaks = y_breaks,
  y_labels = NULL,
  y_axis_labels = y_labels,
  y_max = NULL,
  y_min = NULL,
  expand_y = 0.1,
  expand_x = waiver(),
  censor_fn = rescale_none,
  point_size = 2,
  point_colors = colors,
  band_width = "nrd0",
  n_bins = 50,
  seed = 1234,
  scaling = c("width", "count", "area"),
  method = c("density", "counts"),
  show_sig = TRUE,
  sig_method = "p_by_normality",
  stars = TRUE,
  show_ns = TRUE,
  ns_symbol = "ns",
  sig_bar_nudge = if (y_scale == "log") 0.9 else 0.06,
  sig_star_nudge = if (y_scale == "log") -0.45 else sig_bar_nudge - 0.015,
  sig_text_nudge = if (y_scale == "log") 0.5 else sig_bar_nudge + 0.03,
  step_increase = if (y_scale == "log") 0 else sig_bar_nudge + 0.05,
  sig_bar_thickness = 0.6,
  sig_bar_color = "black",
  sig_font_color = "black",
  sig_star_size = 20,
  sig_text_size = 12,
  p_case = "upper",
  p_spaces = TRUE,
  ...
)
}
\arguments{
\item{df}{Data frame in long format containing continuous variable, categorical or integer variable, and grouping variable}

\item{formula}{y ~ x format}

\item{grouping_var}{Variable to group by. Enter as quoted or unquoted variable name}

\item{grouping_var_order}{Order for levels of \code{grouping_var} along x axis. Enter as character vector}

\item{rev_grouping_var_order}{If \code{TRUE}, levels of \code{grouping_var} are reversed}

\item{x_order}{Order of x values. Enter as character vector. Default is \code{NULL}}

\item{fn_x_order}{Function applied to y to determine order of x. Default is \code{mean}}

\item{rev_x_order}{If \code{FALSE} (default), order of x axis variable is not reversed}

\item{x, y}{Variables for x  and y axis, respectively. Enter as quoted or unquoted variable names}

\item{colors}{Alias for \code{point_colors}}

\item{point_color_var}{Variable used to determine point colors. Enter as quoted or unquoted variable name. Default uses \code{grouping_var}}

\item{point_shape_var}{Variable used to determine point shapes. Default uses \code{grouping_var.} Enter as quoted or unquoted variable name}

\item{point_shapes}{Point shapes. Options: \code{"circle"} (default), \code{"square"}, \code{"triangle"}, \code{"diamond"}. Enter as quoted shape}

\item{alpha}{Alias for \code{point_alpha}}

\item{point_alpha}{Transparency of point colors. Enter as numeric 0-1. Default is \code{0.9}}

\item{width}{Alias for either \code{jitter_width} or \code{beeswarm_width}}

\item{dodge}{Entry for position_dodge. Default is \code{0.7}}

\item{point_border_thickness}{Thickness of line surrounding points in pts units. Default is \code{1}}

\item{show_legend}{If \code{FALSE} (default), legend is not displayed}

\item{legend_title}{Legend title. Default is blank}

\item{y_scale}{Scaling for y axis. Options: \code{"regular"} (default), \code{"log"}, \code{"scientific"}}

\item{y_title}{Alias for \code{y_axis_title}}

\item{y_axis_title}{Enter as quoted string}

\item{x_title}{Alias for \code{x_axis_title}}

\item{x_axis_title}{Enter as quoted string}

\item{plot_title}{Text to be used as main plot title. Enter as quoted string. Default is \code{""} (i.e. no title)}

\item{n_breaks}{Number of tick marks for y axis. Default is \code{3}}

\item{breaks_fn}{Function used to generate y axis breaks. Passed to \code{.create_axis_breaks}. Default is \code{pretty}}

\item{x_breaks}{Numeric vector or function specifying location of ticks along x axis}

\item{x_axis_breaks}{Numeric vector or function specifying location of ticks along x axis}

\item{x_labels}{Vector or function specifying x axis tick labels}

\item{x_axis_labels}{Vector or function specifying x axis tick labels}

\item{y_breaks}{Numeric vector or function specifying location of ticks along y axis}

\item{y_axis_breaks}{Numeric vector or function specifying location of ticks along y axis}

\item{y_labels}{Vector or function specifying y axis tick labels}

\item{y_axis_labels}{Vector or function specifying y axis tick labels}

\item{y_max}{Maximum y axis value. Default is \code{NULL}}

\item{y_min}{Minimum value to use for y axis. Default is \code{0}}

\item{expand_y}{Expansion of y axis around y = 0. Default is \code{0}}

\item{expand_x}{Argument entered for expand argument of \code{scale_x_continuous}. Default is \code{waiver()}}

\item{censor_fn}{Function used to transform data outside y axis limits. Default is \code{rescale_none.} Alternative: \code{scales::censor}}

\item{point_size}{Point size in pts units. Default is \code{5}}

\item{point_colors}{Point fill colors. Enter as character vector of color names or hexadecimal codes}

\item{band_width}{Bandwidth. Options: \code{"nrd0"} (default), \code{"nrd"}, \code{"ucv"}, \code{"bcv"}, \code{"sj"}, \code{"sj-ste"}, \code{"sj-dpi"}, or integer}

\item{n_bins}{Number of bins. Default is \code{50}}

\item{seed}{Number to input into random number generator for reproducibility}

\item{scaling}{Method used to make violin size equal across groups. Options: \code{"width"} (default), \code{"area"}, \code{"count"}}

\item{method}{Method used to make width equal across groups. Options: \code{"density"} (default), \code{"counts"}}

\item{show_sig}{If \code{TRUE} (default), significance annotation displayed. If \code{FALSE}, no significance annotation displayed}

\item{sig_method}{Options: \code{"p_by_normality"} (default), \code{"mann_whitney"}, \code{"t_test_welch"}, \code{"t_test_no_welch"}. Enter as quoted test type}

\item{stars}{If \code{TRUE} (default), significance stars are displayed. If \code{FALSE}, actual P values are displayed}

\item{show_ns}{If \code{TRUE} (default), non-significant annotations are displayed. If \code{FALSE}, no annotation is displayed comparisons in which P >= 0.05}

\item{ns_symbol}{Symbol to use when P >= 0.05. Only relevant when \code{show_ns = TRUE}. Default is \code{"ns"}}

\item{sig_bar_nudge}{Nudge factor along y axis for significance bars. Distance between point or bar and significance bar is sig_bar_nudge x maximum y value on plot. Default is \code{0.08}}

\item{sig_star_nudge}{Nudge factor along y axis for significance stars. Distance between point or bar and significance stars is \code{sig_star_nudge}*maximum y value on plot. Default is \code{0.05}}

\item{sig_text_nudge}{Nudge factor along y axis for significance annotation text. Distance between point or bar and significance annotation text is \code{sig_text_nudge}*maximum y value on plot. Default is \code{0.09}}

\item{step_increase}{Step increase as proportion of view port. Default is \code{0.11}}

\item{sig_bar_thickness}{Line thickness for significance bars in linewidth units. Default is \code{0.6}}

\item{sig_bar_color}{Color for significance bars and annotations. Default is \code{"black"}}

\item{sig_font_color}{Color for text in significance annotations. Default is \code{"black"}}

\item{sig_star_size}{Size of significance stars in pts. Only relevant when \code{stars = TRUE}. Default is \code{20}}

\item{sig_text_size}{Font size for P values in pts. Only relevant when \code{stars = FALSE}. Default is \code{12}}

\item{p_case}{Case of "p" used in significance annotations when \code{stats = FALSE.} Options: \code{"upper"} ("P", default) or \code{"lower"} ("p"). Enter as quoted text}

\item{p_spaces}{If \code{TRUE} (default), spaces placed between p, equality symbol, and P-value in significance annotation. If \code{FALSE}, no space between "p", equality symbol, and P-value in significance annotation. Only relevant when \code{stars = FALSE}}

\item{...}{Arguments passed to \code{theme_fn()}}
}
\value{
ggplot object
}
\description{
Functionality from Thomas Lin Pedersen's excellent package ggforce
}
\examples{
# plot_sina(covid, age ~ gender, y_axis_title = "Age")
# plot_sina(covid, age ~ severity, x_angle = 45, y_axis_title = "Age")

}
