% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_smooth.R
\name{plot_smooth}
\alias{plot_smooth}
\title{Plot regression line or loess smoothed mean per group}
\usage{
plot_smooth(
  df,
  formula = NULL,
  grouping_var = NULL,
  grouping_var_order = NULL,
  rev_grouping_var_order = FALSE,
  y = NULL,
  x = NULL,
  smoothing_fn = stats::loess,
  plot_formula = y ~ x,
  line_type = "solid",
  colors = c("#2A2D34", "#BC3C29", "#00A1D5", "#6761A8", "#009872"),
  line_colors = colors,
  line_color_var = NULL,
  line_thickness = 1,
  show_error = TRUE,
  ci = 0.95,
  error_fill_color = line_colors,
  show_points = FALSE,
  points_in_front = FALSE,
  point_colors = line_colors,
  point_color_var = NULL,
  point_shapes = 21,
  point_shape_var = NULL,
  point_size = 2,
  point_border_thickness = 0,
  point_border_colors = "black",
  point_alpha = 0.9,
  y_axis_title = waiver(),
  y_scale = "regular",
  expand_y = NULL,
  y_axis_labels = NULL,
  y_axis_breaks = NULL,
  y_min = NULL,
  y_max = NULL,
  x_axis_title = waiver(),
  x_axis_breaks = NULL,
  x_range_data = c(-Inf, Inf),
  x_min = NULL,
  x_max = NULL,
  n_breaks = 4,
  breaks_fn = pretty,
  censor_fn = rescale_none,
  plot_title = NULL,
  show_legend = FALSE,
  legend_title = NULL,
  ...
)
}
\arguments{
\item{df}{Data frame in long format containing continuous variable, categorical or integer variable, and grouping variable}

\item{formula}{y ~ x format}

\item{grouping_var}{Variable to group by. Enter as quoted or unquoted variable name}

\item{grouping_var_order}{Order for levels of \code{grouping_var} along x axis. Enter as character vector}

\item{rev_grouping_var_order}{If \code{TRUE}, levels of \code{grouping_var} are reversed}

\item{x, y}{Variables for x  and y axis, respectively. Enter as quoted or unquoted variable names}

\item{smoothing_fn}{Function used for smoothing. Options: \code{"loess"} (default) \code{"lm"}, \code{"glm"}}

\item{plot_formula}{Formula for smoothing function Default is \code{y ~ x}}

\item{line_type}{Type of line. Options include \code{"solid"} (default), \code{"dashed"}, \code{"longdash"}, \code{"twodash"}, \code{"dotdash"}, \code{"dotted"}. Enter as quoted linetype}

\item{colors}{Alias for \code{point_colors}}

\item{line_colors}{Line colors. Hexadecimal or quoted color names. Default is \code{"black"}}

\item{line_thickness}{Line thickness in mm. Default is \code{1}}

\item{show_error}{If \code{TRUE} (default), error estimate for curve displayed as shaded region}

\item{ci}{Confidence interval for error region. Enter as length 1 numeric 0-1. Default is \code{0.95}}

\item{error_fill_color}{Color of shaded error region. Default is same as \code{colors}. Alternative: \code{"grey60"}}

\item{show_points}{If \code{FALSE} (default), points not displayed. If \code{TRUE}, raw data points displayed}

\item{points_in_front}{If \code{FALSE} (default), points are displayed behind smoothed curve}

\item{point_colors}{Point fill colors. Enter as character vector of color names or hexadecimal codes}

\item{point_color_var}{Variable used to determine point colors. Enter as quoted or unquoted variable name. Default uses \code{grouping_var}}

\item{point_shapes}{Point shapes. Options: \code{"circle"} (default), \code{"square"}, \code{"triangle"}, \code{"diamond"}. Enter as quoted shape}

\item{point_shape_var}{Variable used to determine point shapes. Default uses \code{grouping_var.} Enter as quoted or unquoted variable name}

\item{point_size}{Point size in pts units. Default is \code{5}}

\item{point_border_thickness}{Thickness of line surrounding points in pts units. Default is \code{1}}

\item{point_border_colors}{Color of line surrounding points. Default is \code{"black"}}

\item{point_alpha}{Transparency of point colors. Enter as numeric 0-1. Default is \code{0.9}}

\item{y_axis_title}{Enter as quoted string}

\item{y_scale}{Scaling for y axis. Options: \code{"regular"} (default), \code{"log"}, \code{"scientific"}}

\item{expand_y}{Expansion of y axis around y = 0. Default is \code{0}}

\item{y_axis_labels}{Vector or function specifying y axis tick labels}

\item{y_axis_breaks}{Numeric vector or function specifying location of ticks along y axis}

\item{y_min}{Minimum value to use for y axis. Default is \code{0}}

\item{y_max}{Maximum y axis value. Default is \code{NULL}}

\item{x_axis_title}{Enter as quoted string}

\item{x_axis_breaks}{Numeric vector or function specifying location of ticks along x axis}

\item{x_range_data}{Range of x values for input data. Default is -Inf to Inf. Enter as length 2 numeric vector}

\item{x_min, x_max}{Minimum and maximum values to use for x axis}

\item{n_breaks}{Number of tick marks for y axis. Default is \code{3}}

\item{breaks_fn}{Function used to generate y axis breaks. Passed to \code{.create_axis_breaks}. Default is \code{pretty}}

\item{censor_fn}{Function used to transform data outside y axis limits. Default is \code{rescale_none.} Alternative: \code{scales::censor}}

\item{plot_title}{Text to be used as main plot title. Enter as quoted string. Default is \code{""} (i.e. no title)}

\item{show_legend}{If \code{FALSE} (default), legend is not displayed}

\item{legend_title}{Legend title. Default is blank}

\item{...}{Arguments passed to \code{theme_fn()}}
}
\value{
ggplot object
}
\description{
Plot regression line or loess smoothed mean per group
}
