% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spaghetti.R
\name{plot_spag_facet_strata}
\alias{plot_spag_facet_strata}
\title{Faceted spaghetti plot by slope}
\usage{
plot_spag_facet_strata(
  df,
  y,
  color_by = NULL,
  time = NULL,
  date_t0 = NULL,
  date_t1 = NULL,
  id = "id",
  units = "days",
  n_strata = 4,
  min_time = -Inf,
  max_time = Inf,
  colors = c("#005E7A", "#EE1E7A", "#E69F00", "#00C5CD", "#7D5FA7", "#999999", "#333333",
    "#20854E", "#F0E442", "#97572B", "#FBA27D", "#F7F7F7"),
  line_thickness = 0.7,
  alpha = 0.7,
  line_alpha = alpha,
  y_scale = "regular",
  n_breaks = 3,
  y_axis_breaks = NULL,
  y_min = NULL,
  y_max = NULL,
  expand_y = 0.1,
  y_axis_labels = NULL,
  y_axis_title = waiver(),
  x_axis_title = waiver(),
  x_axis_breaks = waiver(),
  n_x_axis_breaks = 6,
  expand_x = 0.05,
  censor_fn = rescale_none,
  min_n_obs = 1,
  show_legend = FALSE,
  plot_title = NULL,
  aspect_ratio = 0.75,
  theme_fn = theme_custom,
  show_facet_labels = TRUE,
  ...
)
}
\arguments{
\item{df}{Data frame in long format with columns for id, y (continuous variable), x (continuous variable)}

\item{y}{Continuous variable used for y axis. Enter as quoted or unquoted variable name}

\item{color_by}{Variable used to color lines. Enter as quoted or unquoted variable name}

\item{time}{Index variable containing time. Enter as quoted or unquoted variable name. Only required if \code{date_t0} and \code{date_t1} are not entered}

\item{date_t0}{Initial date (t = 0) used to calculate time. Enter as quoted or unquoted variable name. If provided, must also provide \code{date_t1}}

\item{date_t1}{Final date used to calculate time. Enter as quoted or unquoted variable name. If provided, must also provide \code{date_t0}}

\item{id}{Variable to identify subjects. Enter as quoted variable name}

\item{units}{Units for time variable. Enter as quoted or unquoted unit. Default is \code{"days"}}

\item{n_strata}{Number of strata. Enter as numeric. Default is \code{4}}

\item{min_time, max_time}{Minimum and maximum values for time. Enter as numeric}

\item{colors}{Alias for \code{line_color}}

\item{line_thickness}{Thickness of line in mm Default is \code{0.7}}

\item{alpha}{Alias for \code{line_alpha}}

\item{line_alpha}{Transparency of lines. Default is \code{0.7}}

\item{y_scale}{Scaling for y-axis. Options: \code{"regular"} (default), \code{"log"}, \code{"scientific"}}

\item{n_breaks}{Number of tick marks for y axis. Default is \code{3}}

\item{y_min}{Minimum value to use for y axis Default is \code{NULL}}

\item{y_max}{Maximum value to use for y axis Default is \code{NULL}}

\item{expand_y}{Expansion of y axis around y = 0. Default is \code{0.1}}

\item{y_axis_labels}{Vector or function specifying y axis tick labels}

\item{x_axis_title, y_axis_title}{Titles for x and y axis, respectively. Default is \code{NULL}}

\item{x_axis_breaks, y_axis_breaks}{Numeric vector or function specifying location of ticks along x and y axis, respectively}

\item{n_x_axis_breaks}{Number of x axis breaks. Only relevant when \code{x_axis_breaks = waiver()}}

\item{expand_x}{Expansion of x axis at ends. Enter as length 1 or 2 (lower, upper) numeric vector. Default is \code{0.05}}

\item{censor_fn}{Function used to transform data outside y axis limits. Default is \code{rescale_none.} Alternative: \code{scales::censor}}

\item{min_n_obs}{Minimum number of observations in each strata. Enter as numeric. Default is \code{1}}

\item{show_legend}{If \code{FALSE} (default), legend is not displayed. If \code{TRUE}, legend for lines is shown}

\item{plot_title}{Text to be used as main plot title. Enter as quoted string}

\item{aspect_ratio}{Aspect ratio. Passed to \code{theme_custom()}. Enter as length 1 numeric. Default is \code{0.75}}

\item{show_facet_labels}{If \code{TRUE} (default), facet labels are displayed}

\item{...}{Arguments passed to \code{theme_custom()}}
}
\value{
Spaghetti plot facetted by slope
}
\description{
Faceted spaghetti plot by slope
}
