% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_summary.R
\name{plot_summary}
\alias{plot_summary}
\title{Plot statistical summary value +/- error of a continuous variable for each group}
\usage{
plot_summary(
  df,
  formula = NULL,
  grouping_var = NULL,
  x = NULL,
  y = NULL,
  dodge = 0.7,
  x_order = NULL,
  rev_x_order = FALSE,
  grouping_var_order = NULL,
  rev_grouping_var_order = FALSE,
  point_shapes = "square",
  point_shape_var = NULL,
  point_color_var = NULL,
  colors = c("#0072B5", "#BC3C29", "#868686", "#CC79A7", "#56B4E9", "#E69F00", "#20854E",
    "#F0E442", "#7D5FA7", "#333333", "#97572B", "#005E7A", "#00A0B0", "#8D9D58",
    "#DA3978", "#61D04F", "#FBA27D", "#FFD500", "#00C5CD", "#F7F7F7"),
  point_colors = colors,
  point_size = 3,
  point_border_thickness = 1,
  point_border_colors = "black",
  show_errorbar = TRUE,
  summary_fn = Mean,
  error_fn = SE,
  errorbar_width = 0.25,
  errorbar_thickness = 0.75,
  show_legend = FALSE,
  legend_title = "",
  y_title = waiver(),
  y_axis_title = y_title,
  y_scale = "regular",
  y_axis_breaks = NULL,
  y_axis_labels = NULL,
  breaks_fn = pretty,
  n_breaks = 4,
  y_min = NULL,
  y_max = NULL,
  expand_y = 0.1,
  censor_fn = rescale_none,
  x_title = NULL,
  x_axis_label_angle = 45,
  x_angle = x_axis_label_angle,
  x_axis_title = x_title,
  x_axis_breaks = NULL,
  x_axis_labels = NULL,
  expand_x = waiver(),
  plot_title = NULL,
  show_sig = FALSE,
  sig_method = "p_by_normality",
  stars = TRUE,
  show_ns = TRUE,
  ns_symbol = "ns",
  sig_bar_nudge = if (y_scale == "log") 0.9 else 0.06,
  sig_star_nudge = if (y_scale == "log") -0.45 else sig_bar_nudge - 0.015,
  sig_text_nudge = if (y_scale == "log") 0.5 else sig_bar_nudge + 0.03,
  step_increase = if (y_scale == "log") 0 else sig_bar_nudge + 0.05,
  sig_bar_thickness = 0.6,
  sig_bar_color = "black",
  sig_font_color = "black",
  sig_star_size = 20,
  sig_text_size = 12,
  p_case = "upper",
  p_spaces = TRUE,
  theme_fn = theme_custom,
  ...
)
}
\arguments{
\item{df}{Data frame in long format containing continuous variable, categorical or integer variable, and grouping variable}

\item{formula}{y ~ x format}

\item{grouping_var}{Variable to group by. Enter as quoted or unquoted variable name}

\item{x, y}{Variables for x  and y axis, respectively. Enter as quoted or unquoted variable names}

\item{dodge}{Entry for position_dodge. Default is \code{0.7}}

\item{x_order}{Order of x values. Enter as character vector. Default is \code{NULL}}

\item{rev_x_order}{If \code{FALSE} (default), order of x axis variable is not reversed}

\item{grouping_var_order}{Order for levels of \code{grouping_var} along x axis. Enter as character vector}

\item{rev_grouping_var_order}{If \code{TRUE}, levels of \code{grouping_var} are reversed}

\item{point_shapes}{Point shapes. Options: \code{"circle"} (default), \code{"square"}, \code{"triangle"}, \code{"diamond"}. Enter as quoted shape}

\item{point_shape_var}{Variable used to determine point shapes. Default uses \code{grouping_var.} Enter as quoted or unquoted variable name}

\item{point_color_var}{Variable used to determine point colors. Enter as quoted or unquoted variable name. Default uses \code{grouping_var}}

\item{colors}{Alias for \code{point_colors}}

\item{point_colors}{Point fill colors. Enter as character vector of color names or hexadecimal codes}

\item{point_size}{Point size in pts units. Default is \code{5}}

\item{point_border_thickness}{Thickness of line surrounding points in pts units. Default is \code{1}}

\item{point_border_colors}{Color of line surrounding points. Default is \code{"black"}}

\item{show_errorbar}{If \code{TRUE}, error bars are displayed}

\item{summary_fn}{Function used to determine midpoint of error bars. Default is \code{Mean}}

\item{error_fn}{Function used to determine ends of error bars. Default is \code{SE}}

\item{errorbar_width}{Width of error bars. Default is \code{n_unique(df$x)/10}}

\item{errorbar_thickness}{Line thickness of error bar. Default is \code{0.75}}

\item{show_legend}{If \code{FALSE} (default), legend is not displayed}

\item{legend_title}{Legend title. Default is blank}

\item{y_title}{Alias for \code{y_axis_title}}

\item{y_axis_title}{Enter as quoted string}

\item{y_scale}{Scaling for y axis. Options: \code{"regular"} (default), \code{"log"}, \code{"scientific"}}

\item{y_axis_breaks}{Numeric vector or function specifying location of ticks along y axis}

\item{y_axis_labels}{Vector or function specifying y axis tick labels}

\item{breaks_fn}{Function used to generate y axis breaks. Passed to \code{.create_axis_breaks}. Default is \code{pretty}}

\item{n_breaks}{Number of tick marks for y axis. Default is \code{3}}

\item{y_min}{Minimum value to use for y axis. Default is \code{0}}

\item{y_max}{Maximum y axis value. Default is \code{NULL}}

\item{expand_y}{Expansion of y axis around y = 0. Default is \code{0}}

\item{censor_fn}{Function used to transform data outside y axis limits. Default is \code{rescale_none.} Alternative: \code{scales::censor}}

\item{x_title}{Alias for \code{x_axis_title}}

\item{x_axis_title}{Enter as quoted string}

\item{x_axis_breaks}{Numeric vector or function specifying location of ticks along x axis}

\item{x_axis_labels}{Vector or function specifying x axis tick labels}

\item{expand_x}{Argument entered for expand argument of \code{scale_x_continuous}. Default is \code{waiver()}}

\item{plot_title}{Text to be used as main plot title. Enter as quoted string. Default is \code{""} (i.e. no title)}

\item{show_sig}{If \code{TRUE} (default), significance annotation displayed. If \code{FALSE}, no significance annotation displayed}

\item{sig_method}{Options: \code{"p_by_normality"} (default), \code{"mann_whitney"}, \code{"t_test_welch"}, \code{"t_test_no_welch"}. Enter as quoted test type}

\item{stars}{If \code{TRUE} (default), significance stars are displayed. If \code{FALSE}, actual P values are displayed}

\item{show_ns}{If \code{TRUE} (default), non-significant annotations are displayed. If \code{FALSE}, no annotation is displayed comparisons in which P >= 0.05}

\item{ns_symbol}{Symbol to use when P >= 0.05. Only relevant when \code{show_ns = TRUE}. Default is \code{"ns"}}

\item{sig_bar_nudge}{Nudge factor along y axis for significance bars. Distance between point or bar and significance bar is sig_bar_nudge x maximum y value on plot. Default is \code{0.08}}

\item{sig_star_nudge}{Nudge factor along y axis for significance stars. Distance between point or bar and significance stars is \code{sig_star_nudge}*maximum y value on plot. Default is \code{0.05}}

\item{sig_text_nudge}{Nudge factor along y axis for significance annotation text. Distance between point or bar and significance annotation text is \code{sig_text_nudge}*maximum y value on plot. Default is \code{0.09}}

\item{step_increase}{Step increase as proportion of view port. Default is \code{0.11}}

\item{sig_bar_thickness}{Line thickness for significance bars in linewidth units. Default is \code{0.6}}

\item{sig_bar_color}{Color for significance bars and annotations. Default is \code{"black"}}

\item{sig_font_color}{Color for text in significance annotations. Default is \code{"black"}}

\item{sig_star_size}{Size of significance stars in pts. Only relevant when \code{stars = TRUE}. Default is \code{20}}

\item{sig_text_size}{Font size for P values in pts. Only relevant when \code{stars = FALSE}. Default is \code{12}}

\item{p_case}{Case of "p" used in significance annotations when \code{stats = FALSE.} Options: \code{"upper"} ("P", default) or \code{"lower"} ("p"). Enter as quoted text}

\item{p_spaces}{If \code{TRUE} (default), spaces placed between p, equality symbol, and P-value in significance annotation. If \code{FALSE}, no space between "p", equality symbol, and P-value in significance annotation. Only relevant when \code{stars = FALSE}}

\item{theme_fn}{Theme function. Default is \code{theme_custom}}

\item{...}{Arguments passed to \code{theme_fn()}}
}
\value{
ggplot object
}
\description{
Plot statistical summary value +/- error of a continuous variable for each group
}
