% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_swimmer.R
\name{plot_swimmer}
\alias{plot_swimmer}
\title{Swimmer plot}
\usage{
plot_swimmer(
  df,
  time_var = "time",
  id = "id",
  point_size = 2,
  point_shape = 21,
  point_color = "#BC3C29",
  point_border_thickness = 0,
  point_border_color = "black",
  line_color = "grey40",
  line_thickness = 0.3,
  show_gridlines = TRUE,
  gridline_thickness = 0.4,
  gridline_color = "grey60",
  gridline_linetype = "dashed",
  x_axis_title = waiver(),
  y_axis_labels = NULL,
  show_y_axis_labels = FALSE,
  show_y_axis_ticks = FALSE,
  expand_x = ggplot2::expansion(c(0.02, 0.02)),
  expand_y = ggplot2::expansion(c(0.02, 0.02)),
  theme = theme_plain(base_size = 14, ...),
  ...
)
}
\arguments{
\item{df}{Data frame in long format with columns for id, y (continuous variable), x (continuous variable)}

\item{time_var}{Column in \code{df} containing time variable. Enter as quoted variable name. Default is \code{"time"}}

\item{id}{Variable to identify subjects. Enter as quoted variable name}

\item{point_size, point_color, point_border_thickness, point_border_color}{Point size, color, border thickness, border color}

\item{line_color}{Line colors. Enter as character vector of hexadecimal codes and/or quoted color names}

\item{line_thickness}{Thickness of line in mm Default is \code{0.7}}

\item{show_gridlines}{If \code{TRUE} (default), grid lines are displayed}

\item{gridline_color, gridline_thickness}{Line color and thickness for grid lines}

\item{y_axis_labels}{Vector or function specifying y axis tick labels}

\item{show_y_axis_labels, show_y_axis_ticks}{If \code{TRUE}, y axis labels or ticks (indicating values of \code{id}) are displayed. Default is \code{FALSE} for both}

\item{expand_x}{Expansion of x axis at ends. Enter as length 1 or 2 (lower, upper) numeric vector. Default is \code{0.05}}

\item{expand_y}{Expansion of y axis around y = 0. Default is \code{0.1}}

\item{theme}{Plot theme}

\item{...}{Arguments passed to \code{theme_custom()}}
}
\value{
ggplot object
}
\description{
Swimmer plot
}
