% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pneumonia_lookup}
\alias{pneumonia_lookup}
\title{Lookup table for pneumonia diagnoses}
\format{
Data frame
}
\usage{
pneumonia_lookup
}
\description{
Lookup table for pneumonia diagnoses
}
\keyword{datasets}
