% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{print.pheatmap}
\alias{print.pheatmap}
\title{Print pheatmap}
\usage{
\method{print}{pheatmap}(x, ...)
}
\arguments{
\item{x}{pheatmap object}

\item{...}{Not used}
}
\value{
pheatmap object with ggplot appended
}
\description{
Print pheatmap
}
