% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_values.R
\name{q_value}
\alias{q_value}
\title{q value}
\usage{
q_value(x)
}
\arguments{
\item{x}{Numeric vector of P values}
}
\value{
Numeric vector with the same length as input
}
\description{
Functionality from \code{qvalue::qvalue}
}
