% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{r}
\alias{r}
\alias{r.default}
\alias{r.factor}
\alias{r.character}
\alias{r.integer}
\alias{r.logical}
\alias{r.Date}
\alias{r.numeric}
\alias{r.list}
\alias{r.data.frame}
\alias{r.grouped_df}
\alias{r.lm}
\alias{r.glm}
\alias{r.coxph}
\alias{r.survdiff}
\alias{r.table}
\alias{r.matrix}
\title{Generic function for rapid display of objects}
\usage{
r(x, ...)

\method{r}{default}(x, ...)

\method{r}{factor}(x, show_plot = TRUE, plot_fn = plot_freq, ...)

\method{r}{character}(x, ...)

\method{r}{integer}(x, ...)

\method{r}{logical}(x, ...)

\method{r}{Date}(
  x,
  month_format = c("number", "name", "abbreviation"),
  leading_zero = FALSE,
  full_year = FALSE,
  date_sep = "-",
  n_as_factor = 10,
  show_plot = TRUE,
  ...
)

\method{r}{numeric}(
  x,
  n_bins = 20,
  color = "#328EC3",
  show_plot = TRUE,
  x_axis_limits = NULL,
  x_axis_breaks = NULL,
  x_axis_title = NULL,
  x_scale = "regular",
  expand_x = if (n_bins > 10) 0.1 else 0.15,
  n_as_factor = 10,
  n_breaks = 6,
  y_axis_title = "n",
  ...
)

\method{r}{list}(x, ...)

\method{r}{data.frame}(x, ...)

\method{r}{grouped_df}(x, ...)

\method{r}{lm}(x, ...)

\method{r}{glm}(x, ...)

\method{r}{coxph}(x, ...)

\method{r}{survdiff}(x, ...)

\method{r}{table}(x, ...)

\method{r}{matrix}(x, ...)
}
\arguments{
\item{x}{Data vector}

\item{...}{Arguments passed to \code{plot_incidence} or \code{plot_fn}}

\item{show_plot}{If \code{TRUE} (default), histogram is plotted}

\item{plot_fn}{Function used for plotting. Default is \code{plot_freq}}

\item{month_format}{Options: \code{"number"} (default), \code{"name"}, \code{"abbreviation"}}

\item{leading_zero}{If \code{FALSE} (default), leading 0s are removed from month and day if < 10}

\item{full_year}{If \code{FALSE} (default), last 2 digits are used. If \code{TRUE}, year is displayed using 4 digits}

\item{date_sep}{Separator between month, day, and year}

\item{n_as_factor}{If number of unique values of \code{x} is < \code{n_as_factor}, \code{x} will be treated as a factor. Default is \code{10}}

\item{n_bins}{Number of bins for x axis. Default is \code{20}}

\item{color}{Color used for histogram bars. Default is \code{"#328EC3"}}

\item{x_axis_limits, x_axis_breaks, x_axis_title}{Limits, breaks, and title for x axis respectively}

\item{x_scale}{Scale for x axis. Default is \code{"regular"}}

\item{expand_x}{Expansion multiplier for x axis. Default is \code{0.1} if \code{n_bins > 10} and 0.15 if \code{n_bins <= 10}}

\item{n_breaks}{Number of desired breaks for x axis ticks. Default is \code{6}}

\item{y_axis_title}{Title for y axis. Default is \code{"n"}}
}
\description{
Generic function for rapid display of objects

r - default

r - factor vector

r - character vector

r - integer vector

r - logical vector

r - date vector

r - numeric vector

r - list

r - data frame

r - grouped_df

r - lm

r - glm

r - coxph

r - survdiff

r - table

r - matrix
}
