% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_descriptive.R
\name{Range}
\alias{Range}
\title{Range}
\usage{
Range(x, finite = TRUE)
}
\arguments{
\item{x}{Numeric or logical vector}

\item{finite}{If \code{TRUE} (default), \code{Inf} (and \code{-Inf}) are removed from \code{x} prior to determining range}
}
\value{
Length 2 numeric vector
}
\description{
Range
}
