% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_rds}
\alias{read_rds}
\alias{read_rda}
\alias{write_rda}
\title{Read rds file}
\usage{
read_rds(file_name, ...)

read_rda(file_name, ...)

write_rda(x, file_name = NULL, directory = desktop_path())
}
\arguments{
\item{file_name}{Name of rda file}

\item{...}{Arguments passed to \code{paste_path()}}

\item{x}{Name of object in global environment to save}

\item{directory}{Directory where rda object will be saved. Default is \code{desktop_path()}}
}
\value{
Contents of rds file

Contents of rda file

rda file silently exported
}
\description{
Functionality from readr package
}
