% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aou_data.R
\name{read_raw_data}
\alias{read_raw_data}
\title{Read CSV file from workspace}
\usage{
read_raw_data(
  file_path,
  type = c("conditions", "person", "surveys", "procedures", "measurements", "drugs",
    "procedures")
)
}
\arguments{
\item{file_path}{Path to CSV file}

\item{type}{Type of data from All of Us. Options: \code{"conditions"} (default), \code{"person"}, \code{"surveys"}, \code{"procedures"}, \code{"measurements"}, \code{"drugs"}, \code{"procedures"}}
}
\description{
Read CSV file from workspace
}
