% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_rds}
\alias{read_rds}
\alias{read_rda}
\alias{write_rda}
\title{Read rds file}
\usage{
read_rds(x)

read_rda(x)

write_rda(x, file_name = NULL, directory = getwd())
}
\arguments{
\item{x}{Name of object in global environment to save}

\item{file_name}{Name of rda file}
}
\value{
Contents of rds file

rda file silently exported
}
\description{
Functionality from readr package
}
