% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_manipulation.R
\name{remove_plot_margins}
\alias{remove_plot_margins}
\alias{remove_plot_legend}
\title{Remove plot margins}
\usage{
remove_plot_margins(x = NULL)

remove_plot_legend(x = NULL)
}
\arguments{
\item{x}{ggplot object. If \code{NULL} (default), enter as \code{plot + remove_plot_margins()} to remove margins}
}
\value{
ggplot object
}
\description{
Remove plot margins

Remove legend
}
