% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_manipulation.R
\name{rename_by_lookup}
\alias{rename_by_lookup}
\title{Rename variables in data frame, list, or vector using a lookup table}
\usage{
rename_by_lookup(x, lookup, old = NULL, new = NULL)
}
\arguments{
\item{x}{Data frame, list, or vector}

\item{lookup}{Lookup data frame or named character vector. If a named character vector, \code{names(lookup)} must match \code{names(x)} (old column names) and values of \code{lookup} are the replacements (new column names). If \code{lookup} is a data frame, must specify column containing old names \code{old} (\code{lookup[[old]]} must match \code{names(x)}) and replacements (\code{lookup[[new]]})}

\item{old, new}{Column names of \code{lookup} containing old/current column names (\code{names(x)}) and corresponding replacement names, respectively. Enter as length 1 character vectors}
}
\value{
Same class as input
}
\description{
Rename variables in data frame, list, or vector using a lookup table
}
