% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_transformation.R
\name{rescale}
\alias{rescale}
\alias{normalize}
\title{Scale values to specified range}
\usage{
rescale(x, to = c(0, 1), from = Range(x))

normalize(x, to = c(0, 1))
}
\arguments{
\item{x}{Numeric vector}

\item{to}{Desired range of output. Enter as length 2 numeric vector. Must be in ascending order. Default is \code{c(0, 1)}}

\item{from}{Starting range. Enter as length 2 numeric vector. Must be in ascending order. Default uses range of \code{x}}
}
\value{
Numeric vector with same length as input
}
\description{
Functionality from \code{scales::rescale}
}
