% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_manipulation.R
\name{resize_plot}
\alias{resize_plot}
\title{Adjust aspect ratio}
\usage{
resize_plot(x = last_plot(), ratio = 1.618034)
}
\arguments{
\item{x}{ggplot object to resize. Default is last plot}

\item{ratio}{New aspect ratio. Defaults to golden ratio}
}
\description{
Adjust aspect ratio
}
