% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programming.R
\name{RM}
\alias{RM}
\title{Remove all objects from global environment}
\usage{
RM(all_names = FALSE, except = ".Random.seed")
}
\arguments{
\item{all_names}{If \code{TRUE}, object names beginning with "." will be removed from global environment. Default is \code{FALSE}}

\item{except}{Objects to keep in global environment. Enter as character vector of object names. Default is \code{".Random.seed"}}
}
\value{
Objects in global environment (except those specified in \code{except}) silently removed
}
\description{
Remove all objects from global environment
}
