% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_manipulation.R
\name{rot_x_labs}
\alias{rot_x_labs}
\title{Rotate x axis labels}
\usage{
rot_x_labs(angle = 45, remove_title = TRUE)
}
\arguments{
\item{angle}{Angle to rotate x axis labels. Default is \code{45}}

\item{remove_title}{If \code{TRUE} (default), x axis title also removed}
}
\value{
ggplot object with x axis labels rotated \code{angle} degrees. Enter as \code{plot + rot_x_labs()}
}
\description{
Rotate x axis labels
}
