% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_axis.R
\name{scale_axis}
\alias{scale_axis}
\alias{scale_y}
\alias{scale_x}
\title{Scale for continuous axis}
\usage{
scale_axis(
  axis = "y",
  scale = "regular",
  title = waiver(),
  limits = NULL,
  breaks = NULL,
  labels = NULL,
  expand_lower = NULL,
  expand_upper = 0,
  expand = c(expand_lower, expand_upper),
  expand_method = c("mult", "add"),
  cap = "upper",
  position = NULL,
  censor_fn = rescale_none,
  transform = NULL,
  ...
)

scale_y(
  axis = "y",
  scale = "regular",
  title = waiver(),
  limits = NULL,
  breaks = NULL,
  labels = NULL,
  expand_lower = NULL,
  expand_upper = 0,
  expand = c(expand_lower, expand_upper),
  expand_method = c("mult", "add"),
  cap = "upper",
  position = NULL,
  censor_fn = rescale_none,
  transform = NULL,
  ...
)

scale_x(
  scale = "regular",
  title = NULL,
  limits = NULL,
  breaks = NULL,
  labels = NULL,
  expand_lower = 0.05,
  expand_upper = expand_lower,
  expand = c(expand_lower, expand_upper),
  expand_method = c("mult", "add"),
  cap = "none",
  position = "bottom",
  censor_fn = rescale_none,
  transform = NULL,
  ...
)
}
\arguments{
\item{axis}{Options: \code{"y"} (default) or \code{"x"}}

\item{scale}{Options: \code{"regular"} (default), \code{"scientific"}, \code{"log10"}, \code{"log2"}, \code{"log"} (same as \code{"log10"}), \code{"log10_1"}}

\item{title}{Axis title}

\item{limits}{Numeric vector of axis limits positions or function to generate limits when passed raw data}

\item{breaks}{Numeric vector of axis break positions or function to generate breaks when passed axis limits}

\item{labels}{Vector or function specifying axis tick labels}

\item{expand_lower, expand_upper}{Expansion to add to ends of axis. Default for \code{expand_lower} is \code{0.05} if \code{expand_method} is \verb{"mult} and \code{0.6} if \code{expand_method} is \code{add}. Default for \code{expand_upper} is \code{0}}

\item{expand}{Alias for \code{c(expand_lower,expand_upper)}}

\item{expand_method}{Method for expanding limits of axes. Options: \code{"mult"} (default. Proportion of entire axis range to add to axis ends), \code{"add"} (add \code{expand_upper} value to upper end and subtract \code{expand_lower} from lower end of axis)}

\item{cap}{Input passed to \code{ggplot2::guide_axis}}

\item{position}{Location of axis. Options: \code{"left"} (default for y axis), \code{"right"}, "bottom" (default for x axis), "top"}

\item{censor_fn}{Function used to transform data outside axis limits. Default is \code{rescale_none.} Alternative: \code{scales::censor}}

\item{...}{Arguments passed to scale function}
}
\value{
ggproto object
}
\description{
Scale for continuous axis

Scale for continuous x axis

Scale for continuous x axis
}
