% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_axis.R
\name{scale_continuous}
\alias{scale_continuous}
\title{Continuous axis scale constructor}
\usage{
scale_continuous(
  axis = "y",
  scale = "regular",
  title = "",
  limits = NULL,
  breaks = NULL,
  labels = NULL,
  expand_lower = NULL,
  expand_upper = 0,
  n_breaks = 4,
  position = NULL,
  censor_fn = rescale_none,
  breaks_fn = pretty,
  cap = "upper",
  ...
)
}
\arguments{
\item{axis}{Options include \code{"y"} (default) or \code{"x"}}

\item{scale}{Options include \code{"regular"} (default), \code{"scientific"}, \code{"log10"}, \code{"log2"}, \code{"log"} (same as \code{"log10"})}

\item{title}{Axis title}

\item{limits}{Minimum and maximum values in data. Must enter in order, i.e. c(lower, upper)}

\item{breaks}{Numeric vector or function specifying location of ticks along axis}

\item{labels}{Vector or function specifying axis tick labels}

\item{expand_lower, expand_upper}{Expansion multiplier around lower and upper limits of axis}

\item{n_breaks}{Desired number of axis breaks. Default is \code{4}}

\item{position}{Location of axis. Options: \code{"left"} (default for y axis), \code{"right"}, "bottom" (default for x axis), "top"}

\item{censor_fn}{Function used to transform data outside axis limits. Default is \code{rescale_none.} Alternative: \code{scales::censor}}

\item{breaks_fn}{Function used to create breaks. Enter as call using (). Default is \code{pretty}. Alternative is \code{breaks_extended}}

\item{cap}{Input passed to \code{ggplot2::guide_axis}}

\item{...}{Arguments passed to scale function}
}
\value{
List passed to \code{ggplot_add.axis_clean}. Enter as \code{plot + scale_continuous()}
}
\description{
Continuous axis scale constructor
}
