% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_subset.R
\name{Select}
\alias{Select}
\alias{remove_constant_cols}
\title{Select columns using tidyselect syntax}
\usage{
Select(.df, ...)

remove_constant_cols(df)
}
\arguments{
\item{...}{Columns to select. Enter using tidyselect syntax}

\item{df}{Data frame}
}
\value{
Data frame
}
\description{
Similar to \code{dplyr::select} but includes all columns when \code{...} are empty
}
