% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_classification.R
\name{sens_spec}
\alias{sens_spec}
\alias{sens_spec.default}
\alias{sens_spec.matrix}
\alias{sens_spec.table}
\alias{sens_spec.data.frame}
\title{Determine sensitivity and specificity at a given cutpoint}
\usage{
sens_spec(
  x,
  outcome_var = NULL,
  predictor_var = NULL,
  cp = NULL,
  type = ">=",
  ci = 0.95,
  yates = FALSE,
  cutpoint = cp,
  n_min_chisq = 5,
  ...
)

\method{sens_spec}{default}(
  x,
  ...,
  outcome_var = NULL,
  predictor_var = NULL,
  cp = NULL,
  type = ">=",
  ci = 0.95,
  yates = FALSE,
  cutpoint = cp,
  n_min_chisq = 5
)

\method{sens_spec}{matrix}(
  x,
  outcome_var = NULL,
  predictor_var = NULL,
  cp = NULL,
  type = ">=",
  ci = 0.95,
  yates = FALSE,
  cutpoint = cp,
  n_min_chisq = 5,
  ...
)

\method{sens_spec}{table}(
  x,
  outcome_var = NULL,
  predictor_var = NULL,
  cp = NULL,
  type = ">=",
  ci = 0.95,
  yates = FALSE,
  cutpoint = cp,
  n_min_chisq = 5,
  ...
)

\method{sens_spec}{data.frame}(
  x,
  outcome_var = NULL,
  predictor_var = NULL,
  cp = NULL,
  type = ">=",
  ci = 0.95,
  yates = FALSE,
  cutpoint = cp,
  n_min_chisq = 5,
  ...
)
}
\arguments{
\item{x}{Several options for input including:
\itemize{
\item Option 1: An integer vector containing counts entered in the following order: test-/control, test-/case, test+/control, test+/case (tn, fn, fp, tp)
\item Option 2: A matrix with columns for test (- in col 1, + in col 2) and rows for truth (- in row 1, + in row 2). tn in row 1/col 1, tp in row 2/col 2. fn in row 2/col 1, fp in row 1, col 2
\item Option 3: A table with structure identical to Option 2
\item Option 4: A data frame
}}

\item{outcome_var}{Binary outcome variable coded as 0 if outcome did not occur and 1 if outcome occurred. Enter as quoted or unquoted variable name. Only relevant when \code{x} is a data frame}

\item{predictor_var}{Continuous predictor variable. Enter as quoted or unquoted variable name. Only relevant when \code{x} is a data frame}

\item{cp, cutpoint}{Cutpoint for binary classification. Will be ignored if entered but \code{predictor_var} is categorical. Only relevant when \code{x} is a data frame}

\item{type}{If \code{>=} (default), positivity is determined by values greater than or equal to cutpoint.  If \code{>}, positivity is determined by values of \code{predictor_var} greater than cutpoint. Only relevant when \code{x} is a data frame}

\item{ci}{Confidence interval. Default is \code{0.95}}

\item{yates}{If \code{TRUE}, Yates' correction is applied to estimation of chi-squared statistic}

\item{n_min_chisq}{Minimum group size that will be allowed for chi-squared test. Default is \code{5}}

\item{...}{Arguments passed to class-specific \code{sens_spec} function}
}
\value{
Data frame with columns "sens" (recall), "sens_lower", "sens_upper", "spec", "spec_lower", "spec_upper", "p", "label", "method", "p_fisher", "p_chi", "youden", "cohen_kappa", "cohen_kappa_lower", "cohen_kappa_upper", "tpr" (power), "fpr" (type 1 error rate), "tnr", "fnr" (type 2 error rate), "fdr", "ppv" (precision), "npv", "lr_pos", "lr_neg", "dor", "dor_lower", "dor_upper", "f1", "accuracy", "auc", "gini", "jaccard_index", "mcc" (Matthews correlation coefficient, phi), "n", "n_outcome", "n_no_outcome", "pos", "neg", "tp", "fn" (type 2 error), "fp" (type 1 error), "tn", "perc_outcome", "perc_no_outcome", "perc_pos", "perc_neg"

Data frame containing the following columns: outcome, predictor, type, cutpoint, sens, sens_lower, sens_upper, spec, spec_lower, spec_upper, p, label, method, p_fisher, p_chi, youden, cohen_kappa, tpr, fpr, tnr, fnr, ppv, npv, fdr, lr_pos, lr_neg, dor, dor_lower, dor_upper, accuracy, auc, gini, jaccard_index, mcc (Matthews correlation coefficient, phi), n, outcome, no_outcome, pos, neg, tp, fn, fp, tn, perc_outcome, perc_no_outcome, perc_pos, perc_neg. Confidence intervals for sensitivity and specificity are approximate binomial
}
\description{
Determine sensitivity and specificity at a given cutpoint

\code{sens_spec} - default method

\code{sens_spec} - matrix

\code{sens_spec} - table

\code{sens_spec} - data frame
}
