% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{set_max_follow_up_time}
\alias{set_max_follow_up_time}
\title{Set maximum duration of follow up time}
\usage{
set_max_follow_up_time(df, max_time, time_var = "time", outcome_var = "death")
}
\arguments{
\item{df}{Data frame}

\item{max_time}{Maximum duration of follow up. Enter as length 1 numeric vector}

\item{time_var}{Variable containing follow up time. Enter as length 1 character vector. Default is \code{"time"}}

\item{outcome_var}{Variable containing outcome variable coded as \code{1} (event) or \code{0} (no event/censored). Enter as length 1 character vector. Default is \code{"death"}}
}
\value{
Data frame with updated time in \code{time_var} column and updated outcome variable in \code{outcome_var} column
}
\description{
Set maximum duration of follow up time
}
