% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{Setdiff}
\alias{Setdiff}
\alias{Intersect}
\alias{Setequal}
\alias{Setequal_all}
\alias{Setdiff_all}
\alias{Setdiff_any}
\alias{Intersect_all}
\title{Elements in x but not y}
\usage{
Setdiff(x, y)

Intersect(x, y)

Setequal(x, y)

Setequal_all(...)

Setdiff_all(...)

Setdiff_any(...)

Intersect_all(...)
}
\arguments{
\item{x, y}{Vectors}

\item{...}{Vectors or list of vectors}
}
\value{
Length 1 logical vector
}
\description{
Similar to \code{base::setdiff} but doesn't convert \code{x} and \code{y} to vectors

Similar to \code{base::intersect} but doesn't convert \code{x} and \code{y} to vectors

Similar to \code{base::setequal} but doesn't convert \code{x} and \code{y} to vectors

Equivalent to sequentially running \code{Setequal} on > 2 vectors

Equivalent to sequentially running \code{Setdiff} on > 2 vectors

Equivalent to combining the results from running \code{Setdiff} and running \code{Setdiff} with input order reversed

Equivalent to sequentially running \code{Intersect} on > 2 vectors
}
