% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{Setequal_all}
\alias{Setequal_all}
\alias{Setdiff_all}
\alias{Setdiff_any}
\alias{Intersect_all}
\title{Setequal on > 2 vectors}
\usage{
Setequal_all(...)

Setdiff_all(...)

Setdiff_any(...)

Intersect_all(...)
}
\arguments{
\item{...}{Vectors or list of vectors}
}
\value{
Length 1 logical vector
}
\description{
Equivalent to sequentially running \code{Setequal} on > 2 vectors

Equivalent to sequentially running \code{Setdiff} on > 2 vectors

Equivalent to combining the results from running \code{Setdiff} and running \code{Setdiff} with input order reversed

Equivalent to sequentially running \code{Intersect} on > 2 vectors
}
