% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_values.R
\name{sig_stars}
\alias{sig_stars}
\title{Convert P values into significance stars}
\usage{
sig_stars(
  x,
  cutpoints = c(0, 1e-04, 0.001, 0.01, 0.05, 1),
  symbols = c("****", "***", "**", "*", "ns")
)
}
\arguments{
\item{x}{Numeric vector of P values}

\item{cutpoints}{Cutpoints for significance stars. Enter as ordered numeric vector that contains 0 as the first value and 1 as the last value. Length should be equal to \code{length(symbols) + 1L}}

\item{symbols}{Symbols representing significance levels from lowest to highest P value. Enter as character vector of length \code{length(cutpoints) - 1L}}
}
\value{
Character vector with same length as input
}
\description{
Convert P values into significance stars
}
