% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{sort_by_name}
\alias{sort_by_name}
\title{Sort vector (or order of column names in a data frame) alphabetically by names}
\usage{
sort_by_name(x, blank_last = TRUE, reverse = FALSE)
}
\arguments{
\item{x}{Named vector, named list, or data frame}

\item{blank_last}{If \code{TRUE} (default), elements without names will be placed at the end}

\item{reverse}{If \code{TRUE}, names ordered reverse alphabetically. Default is \code{FALSE}}
}
\value{
Same class and length as input
}
\description{
Sort vector (or order of column names in a data frame) alphabetically by names
}
