% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{sort_by_nchar}
\alias{sort_by_nchar}
\title{Sort strings by string length (number of characters)}
\usage{
sort_by_nchar(x, desc = FALSE)
}
\arguments{
\item{x}{Vector}

\item{desc}{If \code{FALSE} (default), \code{x} output order is shortest to longest. If \code{TRUE}, output order is longest to shortest}
}
\value{
Same class and length as input
}
\description{
Sort strings by string length (number of characters)
}
