% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_manipulation.R
\name{split_df}
\alias{split_df}
\title{Split data frame by unique values across 1 or more columns}
\usage{
split_df(df, by, sep = "_", drop = FALSE)
}
\arguments{
\item{df}{Data frame}

\item{by}{Columns in \code{df} to split \code{df}. Enter as character, integer, or logical vector}

\item{sep}{Character to separate values. Only relevant if \code{length(by) > 1L}. Enter as length 1 character vector. Default is \code{"_"}}

\item{drop}{Handling of unused levels of grouping variable created by \code{by}. A factor variable is created in which levels (groups) are formed for every possible combination of values across columns defined in \code{by}. \code{drop} is relevant when 1 or more of these potential combinations is not actually present in \code{df}. If \code{drop = TRUE}, such combinations (levels) will be removed from output list. If \code{FALSE} (default), unused combinations will create empty data frames (0 rows) in output list}
}
\value{
List of data frames for each unique combination of values for columns in \code{by}
}
\description{
Split data frame by unique values across 1 or more columns
}
