% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aou_data.R
\name{sql2df}
\alias{sql2df}
\title{Load data from workbench using SQL}
\usage{
sql2df(sql, filename = type, type = NULL)
}
\arguments{
\item{sql}{SQL query. Enter as length 1 character vector}

\item{filename}{Name for output csv file. Enter as length 1 character vector}

\item{type}{Type of data from All of Us. Options: \code{"conditions"}, \code{"person"}, \code{"surveys"}, \code{"procedures"}, \code{"measurements"}, \code{"drugs"}, \code{"procedures"}}
}
\value{
List containing "df" (data frame) and "file" (path to csv file in workspace)
}
\description{
Load data from workbench using SQL
}
