% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aou_sql_strings.R
\docType{data}
\name{sql_bcx}
\alias{sql_bcx}
\title{SQL query to create data frame containing info about BCx}
\format{
An object of class \code{character} of length 1.
}
\usage{
sql_bcx
}
\description{
type: measurement
}
\keyword{datasets}
