% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aou_sql_strings.R
\docType{data}
\name{sql_comorbidities}
\alias{sql_comorbidities}
\title{SQL query to create data frame of cardiometabolic comorbidities}
\format{
An object of class \code{character} of length 1.
}
\usage{
sql_comorbidities
}
\description{
Included comorbidities: CAD, PAD, DM, HLD, hyper-TG, obesity, metabolic syndrome X, PCOS, diabetes insipidus, asthma, COPD, ABPA, liver disease, thyrotoxicosis, goiter
type: condition
}
\keyword{datasets}
