% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aou_sql_strings.R
\docType{data}
\name{sql_demographics}
\alias{sql_demographics}
\title{SQL query to create data frame of patient demographics}
\format{
An object of class \code{character} of length 1.
}
\usage{
sql_demographics
}
\description{
type: survey
}
\keyword{datasets}
