% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aou_sql_strings.R
\docType{data}
\name{sql_inflammatory_markers}
\alias{sql_inflammatory_markers}
\title{SQL query to create data frame containing info about inflammatory markers}
\format{
An object of class \code{character} of length 1.
}
\usage{
sql_inflammatory_markers
}
\description{
type: measurement
}
\keyword{datasets}
