% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aou_sql_strings.R
\docType{data}
\name{sql_lipids_a1c}
\alias{sql_lipids_a1c}
\title{SQL query to create data frame containing info about lipids and HgbA1c}
\format{
An object of class \code{character} of length 1.
}
\usage{
sql_lipids_a1c
}
\description{
type: measurement
}
\keyword{datasets}
