% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aou_sql_strings.R
\docType{data}
\name{sql_patients}
\alias{sql_patients}
\title{SQL query to create data frame of basic patient info with pneumonia}
\format{
An object of class \code{character} of length 1.
}
\usage{
sql_patients
}
\description{
type: person
}
\keyword{datasets}
