% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aou_sql_strings.R
\docType{data}
\name{sql_pneumonia}
\alias{sql_pneumonia}
\title{SQL query to create data frame of patients with pneumonia including info about pneumonia type/date of diagnosis}
\format{
An object of class \code{character} of length 1.
}
\usage{
sql_pneumonia
}
\description{
type: condition
}
\keyword{datasets}
