% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aou_sql_strings.R
\docType{data}
\name{sql_sputum_uag}
\alias{sql_sputum_uag}
\title{SQL query to create data frame containing info about sputum Cx, urine Ag testing, and MRSA nasal swab}
\format{
An object of class \code{character} of length 1.
}
\usage{
sql_sputum_uag
}
\description{
type: measurement
}
\keyword{datasets}
