% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aou_sql_strings.R
\docType{data}
\name{sql_statin}
\alias{sql_statin}
\title{SQL query to create data frame containing info on statin use}
\format{
An object of class \code{character} of length 1.
}
\usage{
sql_statin
}
\description{
type: drugs
}
\keyword{datasets}
