% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aou_sql_strings.R
\docType{data}
\name{sql_substance_use}
\alias{sql_substance_use}
\title{SQL query to create data frame of substance use info}
\format{
An object of class \code{character} of length 1.
}
\usage{
sql_substance_use
}
\description{
type: survey
}
\keyword{datasets}
