% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aou_sql_strings.R
\docType{data}
\name{sql_tn}
\alias{sql_tn}
\title{SQL query to create data frame containing troponin values}
\format{
An object of class \code{character} of length 1.
}
\usage{
sql_tn
}
\description{
type: measurement
}
\keyword{datasets}
