% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_transformation.R
\name{squish}
\alias{squish}
\title{Squish a numeric vector into a specified range}
\usage{
squish(x, min = NULL, max = NULL)
}
\arguments{
\item{x}{Numeric vector}

\item{min, max}{Desired minimum and maximum values. Values in \code{x} below \code{min} will be set to \code{min}. Values in \code{x} above \code{max} will be set to \code{max}. If \code{NULL} (default for both), values outside the limit will not be changed}
}
\value{
Numeric vector with same length as input
}
\description{
Rewritten version of \code{scales::squish}
}
