% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_correlation.R
\docType{data}
\name{stat_cor_anno}
\alias{stat_cor_anno}
\alias{StatCorAnno}
\alias{geom_cor_anno}
\title{Add correlation coefficient and P value to scatter plot}
\format{
An object of class \code{StatCorAnno} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
stat_cor_anno(
  mapping = NULL,
  data = NULL,
  method = "spearman",
  ci = 0.95,
  hypothesis_type = "two.sided",
  auto_label_pos = TRUE,
  label_x_npc = "left",
  label_y_npc = "top",
  label_x = NULL,
  label_y = NULL,
  label_type = "expression",
  coef_format = format_number,
  p_format = format_p_value,
  cor_coef_p_sep = ", ",
  geom = "text",
  position = "identity",
  na.rm = TRUE,
  show.legend = FALSE,
  inherit.aes = TRUE,
  ...
)

StatCorAnno

geom_cor_anno(
  mapping = NULL,
  data = NULL,
  stat = "cor_anno",
  geom = "text",
  position = "identity",
  method = "spearman",
  ci = 0.95,
  hypothesis_type = "two.sided",
  label_type = "expression",
  coef_format = format_number,
  p_format = format_p_value,
  cor_coef_p_sep = ", ",
  auto_label_pos = TRUE,
  label_x_npc = "left",
  label_y_npc = "top",
  label_x = NULL,
  label_y = NULL,
  na.rm = TRUE,
  show.legend = FALSE,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{method}{Correlation test method. Options: \code{"spearman"} (default), \code{"pearson"}, \code{"kendall"}}

\item{ci}{Confidence interval for correlation coefficient. Enter as length 1 numeric 0-1. Default is \code{0.95}}

\item{hypothesis_type}{Alternative hypothesis. Options: \code{"two.sided"} (default), \code{"greater"}, \code{"less"}}

\item{auto_label_pos}{If \code{TRUE} (default), x and y coordinates for annotation automatically determined}

\item{label_x_npc, label_y_npc}{Annotation coordinates in npc units. Can be numeric (between 0-1) or character ("center", "middle", "top", "bottom", "right", "left"). Length should match number of groups/panels, but will be recycled if needed}

\item{label_x, label_y}{Label coordinates in raw data units}

\item{label_type}{Label type. Options: \code{"expression"} (default), \code{"latex"}, \code{"tex"}, \code{"text"}}

\item{coef_format, p_format}{Functions to format correlation coefficient and P value, respectively. Should accept numeric vector as input and return character vector as output}

\item{cor_coef_p_sep}{String that separates correlation coefficient and P value in annotation. Default is \code{", "}}

\item{position}{A position adjustment to use on the data for this layer.
Cannot be jointy specified with \code{nudge_x} or \code{nudge_y}. This
can be used in various ways, including to prevent overplotting and
improving the display. The \code{position} argument accepts the following:
\itemize{
\item The result of calling a position function, such as \code{position_jitter()}.
\item A string nameing the position adjustment. To give the position as a
string, strip the function name of the \code{position_} prefix. For example,
to use \code{position_jitter()}, give the position as \code{"jitter"}.
\item For more information and other ways to specify the position, see the
\link[ggplot2:layer_positions]{layer position} documentation.
}}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{...}{Arguments passed to \code{geom_text} or \code{geom_label}}

\item{stat}{The statistical transformation to use on the data for this layer.
When using a \verb{geom_*()} function to construct a layer, the \code{stat}
argument can be used the override the default coupling between geoms and
stats. The \code{stat} argument accepts the following:
\itemize{
\item A \code{Stat} ggproto subclass, for example \code{StatCount}.
\item A string naming the stat. To give the stat as a string, strip the
function name of the \code{stat_} prefix. For example, to use \code{stat_count()},
give the stat as \code{"count"}.
\item For more information and other ways to specify the stat, see the
\link[ggplot2:layer_stats]{layer stat} documentation.
}}
}
\value{
ggproto object

ggproto object

ggproto object
}
\description{
Functionality from Alboukadel Kassambara's excellent package ggpubr
}
\keyword{datasets}
