% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sig_geoms.R
\docType{data}
\name{StatSig}
\alias{StatSig}
\title{ggproto for Stat to create significance annotation}
\format{
An object of class \code{StatSig} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 6.
}
\usage{
StatSig
}
\description{
ggproto for Stat to create significance annotation
}
\keyword{datasets}
