% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_capitalize}
\alias{str_capitalize}
\alias{str_capitalize_words}
\title{Capitalize first letter of each string in character vector}
\usage{
str_capitalize(x, other_letters_lower = FALSE)

str_capitalize_words(x, other_letters_lower = FALSE)
}
\arguments{
\item{x}{Character vector}

\item{other_letters_lower}{If \code{FALSE} (default), subsequent letters left as is. If \code{TRUE}, subsequent letters converted to lower case}
}
\value{
Character vector with same length as input
}
\description{
Inspired by github.com/raredd/regex
}
