% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{can_be_numeric}
\alias{can_be_numeric}
\alias{str_count}
\title{Is string coercible to numeric}
\usage{
can_be_numeric(x)

str_count(x, pattern, ignore_case = FALSE, ...)
}
\value{
Logical vector with same length as input (unless \code{x} is logical in which case, output is \code{FALSE})

Integer vector with same length as input
}
\description{
Is string coercible to numeric

Count number of occurrences of substring
}
