% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_numerics}
\alias{str_numerics}
\alias{str_nth_number}
\title{Extract all numbers from each element of a character vector}
\usage{
str_numerics(
  x,
  single_number = FALSE,
  incl_decimal = TRUE,
  incl_comma = FALSE,
  incl_neg = FALSE,
  as_numeric = FALSE
)

str_nth_number(
  x,
  n = 1,
  incl_decimal = TRUE,
  incl_comma = FALSE,
  incl_neg = FALSE,
  as_numeric = FALSE
)
}
\arguments{
\item{x}{Character vector}

\item{single_number}{If \code{FALSE} (default), output is a list (same length as input) of numbers. If \code{TRUE}, all numbers in a string are pasted together using \code{paste0()}}

\item{incl_decimal}{If \code{TRUE} (default), decimals can be included in numbers. If \code{FALSE}, values will be split at decimals. Only relevant when \code{single_number = FALSE}}

\item{incl_comma}{If \code{FALSE} (default), values will be split at commas. If \code{TRUE}, commas can be included in numbers. Only relevant when \code{single_number = FALSE}}

\item{incl_neg}{If \code{TRUE}, negative sign can be included in output. If \code{FALSE} (default), negative sign not included in output. Only relevant when \code{single_number = FALSE}}

\item{as_numeric}{If \code{FALSE} (default), output is a character. If \code{TRUE}, output is numeric}

\item{n}{nth occurrence of a number in each element of \code{x}. Options: numeric, \code{"first"} (equivalent to \code{n = 1}), \code{"last"} (equivalent to \code{n = -1}). For second to last number, enter \code{n = -2}}
}
\value{
If \code{single_number = TRUE}, output is a vector (character vector if \code{as_numeric = FALSE}, numeric vector if \code{as_numeric = TRUE}). If \code{single_number = FALSE}, output is a list of vectors, each containing all of the individual numbers in a given string (character vector if \code{as_numeric = FALSE}, numeric vector if \code{as_numeric = TRUE})

Vector (character vector if \code{as_numeric = FALSE}, numeric vector if \code{as_numeric = TRUE}) with same length as input
}
\description{
Extract all numbers from each element of a character vector

Extract nth number from string
}
