% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_remove}
\alias{str_remove}
\alias{str_remove_numerics}
\alias{str_remove_letters}
\alias{str_remove_punc}
\alias{str_trimws}
\alias{str_remove_space}
\alias{str_remove_multi_space}
\alias{str_remove_before}
\alias{str_remove_after}
\title{Remove a pattern from a string}
\usage{
str_remove(x, pattern, all = TRUE, ...)

str_remove_numerics(x, all = TRUE)

str_remove_letters(x, case = "both", all = TRUE)

str_remove_punc(x)

str_trimws(x, side = c("both", "left", "right"))

str_remove_space(x)

str_remove_multi_space(x)

str_remove_before(x, pattern, fixed = TRUE, all = FALSE)

str_remove_after(x, pattern, fixed = TRUE, all = FALSE)
}
\arguments{
\item{x}{Character vector}

\item{pattern}{Pattern to remove from \code{x}. Enter as length 1 character vector}

\item{all}{If \code{TRUE} (default), all occurrences of \code{pattern} will be removed. If \code{FALSE}, only first occurrence of \code{pattern} will be removed}

\item{...}{Arguments passed to \code{sub} or \code{gsub}}

\item{case}{Case of letters to be removed. Options: \code{"both"} (default), \code{"lower"}, \code{"upper"}}

\item{side}{Side of string to remove spaces. Options: \code{"both"} (default), \code{"left"}, \code{"right"}}

\item{fixed}{If \code{TRUE} (default), \code{escape_regex} will be applied to \code{pattern} before running \code{sub} or \code{gsub}}

\item{name}{description}
}
\value{
Character vector with same length as input
}
\description{
Remove a pattern from a string

Remove numbers from a string

Remove letters from string

Alias for \code{str_alphanum}

Wrapper for trimws

Remove all spaces (not just at ends of string)

Convert multiple spaces to single space

Remove substring before a given pattern (i.e. pattern retained in output)

Remove substring before a given pattern (i.e. pattern retained in output)
}
