% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_subset.R
\name{Subset}
\alias{Subset}
\title{Alternative version of \code{base::subset}}
\usage{
Subset(df, cols, rows)
}
\arguments{
\item{df}{Data frame}

\item{cols}{Columns to select. Enter as quoted or unquoted column names or column numbers. For any input type, use \code{c()} if multiple columns are selected. Use \code{-} to exclude columns. Default selects all columns}

\item{rows}{Rows to select. Enter as predicate statement or row numbers. Not compatible with tidyselect input. Default uses all rows}
}
\value{
Data frame with columns specified by cols argument and rows specified by rows input
}
\description{
Similar to using \code{dplyr::select} followed by \code{dplyr::filter}
}
