% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{summary_table_binary}
\alias{summary_table_binary}
\title{Summarize all binary variables stratified by 1 binary grouping variable}
\usage{
summary_table_binary(
  df,
  grouping_var = NULL,
  binary_variables = NULL,
  n_format = "n",
  group_levels = NULL,
  digits_or = 2,
  digits_perc = 1,
  incl_denominator = TRUE,
  perc_first = TRUE,
  n_min_chisq = 5,
  ci = 0.95,
  total_col = TRUE
)
}
\arguments{
\item{df}{Data frame}

\item{grouping_var}{Binary grouping variable used to define columns in output table. Enter as quoted variable name}

\item{binary_variables}{Binary variables to be compared between groups. Enter as comma separated list of quoted variable names wrapped in \code{c()}}

\item{n_format}{Format for sample size entry in column names. Default is \code{n}}

\item{group_levels}{Levels for \code{grouping_var} Enter as length 2 character vector with lower level 1st}

\item{digits_or}{Number of digits used to print odds ratios. Default is \code{2}}

\item{digits_perc}{Number of digits for percentages. Default is \code{1}}

\item{incl_denominator}{If \code{TRUE} (default), format for count is n/total. If \code{FALSE}, format is n}

\item{perc_first}{If \code{TRUE} (default), format is \% (n) or \% (n/n_total). If \code{FALSE}, format is n (\%) or n/total (\%)}

\item{n_min_chisq}{Minimum group size that will be allowed for Chi-squared test. Default is \code{5}}

\item{ci}{Confidence interval. Default is \code{0.95}}

\item{total_col}{If \code{TRUE} (default), column for totals included in output table}
}
\value{
Data frame
}
\description{
Summarize all binary variables stratified by 1 binary grouping variable
}
