% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programming.R
\name{suppress}
\alias{suppress}
\title{Suppress messages and warnings}
\usage{
suppress(...)
}
\arguments{
\item{...}{Expression to evaluate}
}
\value{
Expression entered into \code{...} run without messages or warnings
}
\description{
Suppress messages and warnings
}
