% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{surv_pairwise}
\alias{surv_pairwise}
\title{Pairwise HR, CI, P values}
\usage{
surv_pairwise(
  df,
  predictor_var,
  time_var = "time",
  outcome_var = "death",
  adjust_method = "holm"
)
}
\arguments{
\item{df}{Data frame}

\item{predictor_var}{Column in \code{df} containing grouping variable. Pairwise comparisons will be made across unique values of \code{df[[predictor_var]]}. Enter as character vector}

\item{time_var}{Column in \code{df} containing follow up time. Enter as length 1 character vector. Default is \code{"time"}}

\item{outcome_var}{Column in \code{df} containing outcome variable coded as 1 (event) or 0 (no event/censored). Enter as length 1 character vector. Default is \code{"death"}}

\item{adjust_method}{Method of adjustment for multiple comparisons. Default is \code{"holm"}}
}
\value{
Data frame with columns "Group1", "Group2", "hr", "hr_lower", "hr_upper", "p_lrt", "p_lrt_adj", "label", "p_wald", "p_wald_adj", "n", "n_events"
}
\description{
Pairwise HR, CI, P values
}
