% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_theme.R
\name{theme_facs}
\alias{theme_facs}
\title{Plot theme for FACS data}
\usage{
theme_facs(
  base_size = 14,
  boxed = TRUE,
  axis_line_thickness = 0.4,
  aspect_ratio = 1,
  y_axis_title_angle = 90,
  x_axis_label_angle = 0,
  x_angle = x_axis_label_angle,
  x_axis_label_hjust = 0.5,
  x_axis_label_vjust = 0.5,
  x_axis_label_margin_right = 0,
  x_axis_title_font_size = base_size + 2,
  y_axis_title_font_size = base_size + 2,
  plot_margin = ggplot2::margin(t = 10, r = 20),
  facet_fill_color = "#D9D9D9",
  facet_line_color = "black",
  facet_line_type = 1,
  facet_line_thickness = axis_line_thickness,
  facet_text_size = base_size,
  facet_text_color = NULL,
  facet_text_face = "plain",
  facet_text_margin = ggplot2::margin(t = 5, b = 5),
  ...
)
}
\arguments{
\item{base_size}{Size of text in pts. Default is \code{16}}

\item{boxed}{If \code{TRUE} (default), rectangle drawn around plotting area}

\item{axis_line_thickness}{Axis and tick thickness in mm. Default is \code{0.7}}

\item{aspect_ratio}{Aspect ratio, y/x. Default is \code{1}}

\item{y_axis_title_angle}{Angle of text used for y axis title. Default is 90}

\item{x_axis_label_angle}{Angle for x axis tick labels}

\item{x_angle}{Alias for \code{x_axis_label_angle}}

\item{x_axis_label_hjust, x_axis_label_vjust}{Horizontal and vertical justification of x axis text. Default is \code{0.5}}

\item{x_axis_label_margin_right}{Margin to the right of x axis tick labels. Default is \code{0}}

\item{x_axis_title_font_size, y_axis_title_font_size}{Default is \code{base_size + 2}}

\item{plot_margin}{Plot margin. Enter as \code{margin()}. Default is \code{margin(t = 10, r = 20)}}

\item{facet_fill_color}{Color of facet boxes. Default is \code{"#D9D9D9"}}

\item{facet_line_color}{Color of facet box border. Default is \code{"black"}}

\item{facet_line_type}{Line type of facet box border. Default is \code{1} (solid line)}

\item{facet_line_thickness}{Line thickness of facet box border. Default uses \code{axis_line_thickness}}

\item{facet_text_size, facet_text_color, facet_text_face}{Text style for facet labels}

\item{facet_text_margin}{Margin around facet labels. Enter as \code{margin()}. Default is \code{margin(t = 5, b = 5)}}

\item{...}{Arguments passed to \code{theme()}}
}
\description{
Slightly modified version of \code{theme_plain}
}
