% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{top}
\alias{top}
\title{Display specified number of rows of data frame as tibble}
\usage{
top(df, n = 15)
}
\arguments{
\item{df}{Data frame}

\item{n}{Number of rows to display. Default is \code{15}}
}
\value{
Print first \code{n} rows of \code{df}
}
\description{
Display specified number of rows of data frame as tibble
}
