% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aou_df_load.R
\name{transfer_to_bucket}
\alias{transfer_to_bucket}
\title{Move object from workspace to bucket}
\usage{
transfer_to_bucket(x, filename = substitute(x))
}
\arguments{
\item{x}{Object in global environment}

\item{filename}{Name for exported Rdata file}
}
\value{
Rdata file silently exported
}
\description{
Move object from workspace to bucket
}
