% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_manipulation.R, R/stats_transformation.R
\name{transform_rowwise}
\alias{transform_rowwise}
\alias{normalize_rowwise}
\alias{normalize_colwise}
\title{Row-wise transformation}
\usage{
transform_rowwise(.df, .fn, ...)

normalize_rowwise(df, ..., to = c(0, 1))

normalize_colwise(df, ..., to = c(0, 1))
}
\arguments{
\item{.df}{Data frame}

\item{.fn}{Function used to transform row values. Must be vectorized}

\item{...}{Columns to include in normalization. Enter using tidyselect syntax}

\item{df}{Data frame}

\item{to}{Desired range for output. Enter as length 2 numeric vector in ascending order. Default is \code{c(0, 1)}}
}
\value{
Data frame with same dimensions as input

Data frame with same dimensions as input
}
\description{
Row-wise transformation

Normalize row-wise

Normalize column-wise
}
