% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{tsibble}
\alias{tsibble}
\title{Create time series data frame}
\usage{
tsibble(
  df,
  t0 = NULL,
  t1 = NULL,
  time = NULL,
  id = "id",
  units = c("days", "weeks", "months", "years", "none"),
  min_time = -Inf,
  max_time = Inf,
  new_time_col = "time"
)
}
\arguments{
\item{df}{Data frame}

\item{t0, t1}{Dates for beginning and end of time period, respectively. Enter as length 1 character vector. If provided, must also provide \code{t1}}

\item{time}{Index variable containing time. Enter as length 1 character vector. Only required if \code{t0} and \code{t1} are not specified}

\item{id}{Key variable containing subject/individuals. Enter as length 1 character vector. Default is \code{"id"}}

\item{units}{Units for time variable in output. Enter as length 1 character vector. Options: \code{"days"} (default), \code{"weeks"}, \code{"months"}, \code{"years"}, \code{"none"}}

\item{min_time, max_time}{Minimum and maximum values for time. Enter as length 1 numeric vector with units matching \code{units} argument}

\item{new_time_col}{Name for new time variable. Enter as length 1 character vector}
}
\value{
tsibble object
}
\description{
Create time series data frame
}
