% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectors.R
\name{type_convert}
\alias{type_convert}
\title{Convert a vector to the most likely class intended}
\usage{
type_convert(x)
}
\arguments{
\item{x}{Vector}
}
\value{
Vector of most likely intended class
}
\description{
Convert a vector to the most likely class intended
}
