% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_info.R
\name{vars_which}
\alias{vars_which}
\alias{vars_binary}
\alias{vars_binary_01}
\alias{vars_cat}
\alias{vars_fct}
\alias{vars_chr}
\alias{vars_dbl}
\alias{vars_numeric}
\alias{vars_continuous}
\alias{vars_int}
\alias{vars_can_be_numeric}
\alias{vars_lgl}
\alias{vars_date}
\alias{vars_id}
\alias{vars_constant}
\alias{vars_any_na}
\alias{vars_no_na}
\alias{vars_all_na}
\alias{vars_contain}
\title{Get names of variables based on predicate function applied to column values}
\usage{
vars_which(.df, .predicate_fn, ..., .invert = FALSE)

vars_binary(df, invert = FALSE)

vars_binary_01(df, invert = FALSE)

vars_cat(df, max_n_unique = 10, invert = FALSE)

vars_fct(df, invert = FALSE)

vars_chr(df, invert = FALSE)

vars_dbl(df, invert = FALSE, incl_integer = TRUE)

vars_numeric(df, invert = FALSE, incl_integer = TRUE)

vars_continuous(df, invert = FALSE)

vars_int(df, invert = FALSE)

vars_can_be_numeric(df, invert = FALSE, incl_integer = TRUE)

vars_lgl(df, invert = FALSE)

vars_date(df, invert = FALSE)

vars_id(df, na.rm = FALSE, invert = FALSE)

vars_constant(df, invert = FALSE)

vars_any_na(df, invert = FALSE)

vars_no_na(df, invert = FALSE)

vars_all_na(df, invert = FALSE)

vars_contain(
  df,
  pattern,
  ignore.case = FALSE,
  fixed = FALSE,
  perl = FALSE,
  invert = FALSE
)
}
\arguments{
\item{.df}{Data frame}

\item{.predicate_fn}{Predicate function that will be applied to column values to select variable names. Must evaluate to logical (not \code{NULL}) and should not be vectorized (i.e. for vector input, single value should be returned)}

\item{...}{Arguments passed to \code{.predicate_fn}}

\item{.invert}{If \code{TRUE}, output includes variables for which the predicate is \code{FALSE}}

\item{max_n_unique}{Maximum number of unique values, below which variable assumed to be categorical}

\item{incl_integer}{If \code{TRUE} (default), integer variables are included}

\item{na.rm}{If \code{TRUE}, missing values are not considered}

\item{pattern}{Search pattern to detect in each column of \code{df}. Enter as length 1 character vector}

\item{ignore.case}{If \code{TRUE}, case-insensitive search is performed. Default is \code{FALSE}}

\item{fixed}{If \code{TRUE}, exact match search for \code{pattern} is performed. Default is \code{FALSE}}

\item{perl}{If \code{TRUE}, perl-compatible syntax is allowed for \code{pattern} input. Default is \code{FALSE}}
}
\value{
Character vector containing variable names that result in \code{TRUE} when \code{.predicate_fn} is applied to columns in \code{df}
}
\description{
Get names of variables based on predicate function applied to column values

Get names of binary variables

Get names of binary variables coded as 0 or 1

Get names of categorical variables

Get names of factor variables

Get names of character variables

Get names of numeric variables

Alias for vars_dbl

Get names of continuous variables

Get names of integer variables

Get names of variables which are coercible to a numeric

Get names of logical variables

Get Names of date variables

Get names of id variables

Get names of constant (zero-variance) variables

Get names of variables that contain one or more missing values

Get names of variables that contain no missing values

Get names of variables that contain only missing values

Get names of variables that contain specified substring
}
