% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aou_data.R
\name{write_raw_data}
\alias{write_raw_data}
\title{Using SQL, create data frame and export to workspace}
\usage{
write_raw_data(
  sql,
  filename,
  dir = NULL,
  type = NULL,
  workspace = Sys.getenv("WORKSPACE_CDR"),
  billing = Sys.getenv("GOOGLE_PROJECT"),
  destination_format = "CSV",
  return_df = TRUE
)
}
\arguments{
\item{sql}{String containing SQL query}

\item{filename}{String containing name of file}

\item{dir}{Path to directory where \code{filename} will be stored}

\item{type}{Type of sql query. If \code{NULL} (default), type will be guessed from \code{sql}. Only relevant when \code{return_df = TRUE}}

\item{workspace, billing}{Workspace and billing specific to user's environment}

\item{destination_format}{File type for exported data. Default is \code{"CSV"}}

\item{return_df}{If \code{TRUE} (default), data frame is returned}
}
\value{
Data silently exported. If \code{return_df = TRUE}, output is a data frame, otherwise output is path to file
}
\description{
Using SQL, create data frame and export to workspace
}
