% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_coerce.R
\name{cbind_list}
\alias{cbind_list}
\title{Combine list of vectors column-wise to create data frame}
\usage{
cbind_list(x)
}
\arguments{
\item{x}{List of vectors. Each vector must have same length or must share a common multiple (see examples for details). To convert a list of vectors to rows of a data frame, use \code{rbind_list()}}
}
\value{
Data frame with elements of input list (vectors) forming columns of output data frame
}
\description{
Opposite of \code{as.list.data.frame}
}
\examples{
\dontrun{
## Identical to data.frame(a = 1:3, b = 4:6)
# cbind_list(list(a = 1:3, b = 4:6))
## Identical to data.frame(a = 1:3, b = 4:9)
# cbind_list(list(a = 1:3, b = 4:9))
## Results in error: same as data.frame(a = 1:3, b = 4:8)
# cbind_list(list(a = 1:3, b = 4:8))
}

}
