% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{install_if_missing_then_load}
\alias{install_if_missing_then_load}
\alias{Lib}
\title{Install and load package from cran}
\usage{
install_if_missing_then_load(package, ...)

Lib(
  ...,
  lib.loc = NULL,
  warn.conflicts,
  mask.ok,
  exclude,
  include.only,
  attach.required = missing(include.only)
)
}
\arguments{
\item{package}{Name of package. Enter as length 1  character vector}

\item{...}{Arguments passed to \code{require}}
}
\value{
Packages silently loaded
}
\description{
Install and load package from cran

Load packages from cran
}
