% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_descriptive.R
\name{Quantile}
\alias{Quantile}
\alias{Q1}
\alias{Q3}
\alias{iqr}
\title{Quantile}
\usage{
Quantile(x, probs = c(0, 0.25, 0.5, 0.75, 1), prism = FALSE)

Q1(x)

Q3(x)

iqr(x)
}
\arguments{
\item{x}{Numeric vector}

\item{probs}{Quantiles. Enter as numeric 0-1. Default uses min, Q1, median, Q3, max}

\item{prism}{If \code{TRUE}, calculation performed using same algorithm used by Prism. Default is \code{FALSE}}
}
\value{
Numeric vector with length equal to the length of \code{probs}

Length 2 numeric vector containing 25th and 75th percentile values
}
\description{
Functionality from \code{stats::quantile}
}
