% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_manipulation.R
\name{remove_plot_axis}
\alias{remove_plot_axis}
\title{Remove axis}
\usage{
remove_plot_axis(
  x = NULL,
  axis = c("x", "y"),
  component = c("lines", "ticks", "text", "title")
)
}
\arguments{
\item{x}{ggplot object. If \code{NULL} (default), enter as \code{plot + remove_plot_axis()} to remove axis components}

\item{axis}{Options: \code{"x"}, \code{"y"}, or both (default)}

\item{component}{Components of axis to remove. Options: \code{"lines"}, \code{"ticks"}, \code{"text"}, \code{"title"}. Default includes all. Enter as character vector. Enter \code{NULL} to keep all elements}
}
\value{
theme object with axis removed
}
\description{
Remove axis
}
