% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aou_save_data.R
\name{write_raw_data}
\alias{write_raw_data}
\title{Using SQL, create data frame and export to workspace}
\usage{
write_raw_data(
  sql,
  filename,
  workspace = Sys.getenv("WORKSPACE_CDR"),
  bucket = Sys.getenv("WORKSPACE_BUCKET"),
  email = Sys.getenv("OWNER_EMAIL"),
  billing = Sys.getenv("GOOGLE_PROJECT"),
  destination_format = "CSV"
)
}
\arguments{
\item{sql}{String containing SQL query}

\item{filename}{String containing name of file}

\item{workspace, bucket, email, billing}{Workspace, bucket, email, and billing specific to users environment}

\item{destination_format}{File type for exported data. Default is \code{"CSV"}}

\item{dir}{Path to directory where \code{filename} will be stored}

\item{return_df}{If \code{TRUE} (default), data frame is returned}
}
\value{
Data silently exported. If \code{return_df = TRUE}, output is a data frame, otherwise output is path to file
}
\description{
Using SQL, create data frame and export to workspace
}
